Examinees

Psychometrics.Examinee — Type

Examinee <: AbstractExaminee

Description

Examinee struct with a generic latent variable.

Fields

idx::Int64: An integer that identifies the Examinee in this session.

id::String: A string that identifies the Examinee.

latent::Latent: A generic latent variable associated with the Examinee.

Factories

Examinee(idx, id, latent) = new(idx, id, latent)

Creates a new Examinee with custom index, id and a generic latent variable.

Random initializers

Examinee(idx, id) = new(idx, id, Latent1D())

Randomly generates an Examinee with custom index and id and with a default 1-dimensional latent variable (Look at (Latent1D)[#Psychometrics.Latent1D] for the defaults).

Psychometrics.empty_chain! — Method

empty_chain!(examinee::AbstractExaminee)

Psychometrics.get_examinee_by_id — Method

get_examinee_by_id(examinee_id::String, examinees::Vector{<:AbstractExaminee})

It returns the Examinee with index examinee_id from a Vector of AbstractExaminee.

Psychometrics.get_latents — Method

get_latents(examinees::Vector{<:AbstractExaminee})

Returns a matrix with latent values displayed by row.

Psychometrics.Latent — Type

Latent <: AbstractLatent

Description

Generic latent variable struct.

Fields

- **`val::Vector{Float64}`**
- **`bounds::Vector{Vector{Float64}}`**
- **`prior::Union{Distributions.ContinuousUnivariateDistribution, Distributions.MultivariateDistribution}`**
- **`posterior::Union{Distributions.ContinuousUnivariateDistribution, Distributions.MultivariateDistribution}`**
- **`chain::Vector{Vector{Float64}} `**
- **`expected_information::Matrix{Float64}`**

Factories

Latent(names, val, bounds, prior, posterior, chain, expected_information) = new(val, bounds, prior, posterior, chain, expected_information)

Creates a new generic latent variable with custom fields.

Random Initializers

Latent(bounds, posterior)

Randomly generates a value for the generic latent variable and assigns a custom univariate or multivariate distribution to prior and posterior with specific bounds.

Psychometrics.Latent1D — Type

Latent1D <: AbstractLatent

Description

1-dimensional latent variable struct.

Fields

- **`val::Float64`**
- **`bounds::Vector{Float64}`**
- **`prior::Distributions.ContinuousUnivariateDistribution`**
- **`posterior::Distributions.ContinuousUnivariateDistribution`**
- **`chain::Vector{Float64}`**
- **`expected_information::Float64`**

Factories

Latent1D(val, bounds, prior, posterior, chain, expected_information) = new(val, bounds, prior, posterior, chain, expected_information)

Creates a new 1-dimensional latent variable with custom fields.

Random Initializers

Latent1D()

Randomly generates a value for the 1-dimensional latent variable and assigns a default standardized Gaussian prior and posterior Latent1D(dist, bounds)

Same as Latent1D(dist, bounds) but with value equal to val Latent1D(val)

Randomly generates a value for the 1-dimensional latent variable and assigns a custom univariate distribution to prior and posterior with specific bounds.

Psychometrics._chain_append! — Method

_chain_append!(latent::Latent1D; sampling = false)

Psychometrics._empty_chain! — Method

_empty_chain!(latent::Latent1D)

Psychometrics._set_val! — Method

_set_val!(latent::Latent1D, val::Float64)

Psychometrics._set_val_from_chain! — Method

_set_val_from_chain!(latent::Latent1D)

Psychometrics._update_estimate! — Method

_update_estimate!(latent::Latent1D; sampling = true)

Psychometrics.LatentND — Type

LatentND <: AbstractLatent

Description

N-dimensional latent variable struct.

Fields

names::Vector{String}

val::Vector{Float64}

bounds::Vector{Vector{Float64}}

prior::Distributions.MultivariateDistribution

posterior::Distributions.MultivariateDistribution

chain::Vector{Vector{Float64}}

expected_information::Matrix{Float64}

Factories

LatentND(names, val, bounds, prior, posterior, chain, expectedinformation) = new(names, val, bounds, prior, posterior, chain, expectedinformation) Creates a new N-dimensional latent variable with custom fields.

Random Initializers

LatentND()

Randomly generates a value for the N-dimensional latent variable and assigns a default standardized Gaussian prior and posterior LatentND(dist, bounds)

Randomly generates a value for the N-dimensional latent variable and assigns a custom multivariate distribution to prior and posterior with specific bounds.