Fisher Information

Psychometrics.expected_information_item — Method

expected_information_item(examinees::Vector{<:AbstractExaminee}, items::Vector{<:AbstractItem})

Description

Abstraction of expectedinformationitem(latent, parameters) on Vector{<:AbstractExaminee} and items::Vector{<:AbstractItem}. It follows the parametrization $a(θ - b)$.

Arguments

examinee::Vector{<:AbstractExaminee} : Required.

item::AbstractItem : Required.

Output

A matrix (or a scalar if there is only on item parameter) of the expected informations.

Psychometrics.expected_information_item — Method

expected_information_item(latent::Latent1D, parameters::Parameters1PL)

Description

It computes the expected information (-second derivative of the likelihood) with respect to the difficulty parameter of the 1PL model. It follows the parametrization $a(θ - b)$.

Arguments

latent::Latent1D : Required. A 1-dimensional Latent1D latent variable.

parameters::Parameters1PL : Required. A 1-parameter logistic parameters object.

Output

A Float64 scalar.

Psychometrics.expected_information_item — Method

expected_information_item(latent::Latent1D, parameters::Parameters2PL)

Description

It computes the expected information (-second derivative of the likelihood) with respect to the 2 parameters of the 2PL model. It follows the parametrization $a(θ - b)$.

Arguments

latent::Latent1D : Required. A 1-dimensional Latent1D latent variable.

parameters::Parameters1PL : Required. A 2-parameter logistic parameters object.

Output

A $2 \times 2$ matrix of the expected informations.

Psychometrics.expected_information_item — Method

expected_information_item(latent::Latent1D, parameters::Parameters3PL)

Description

It computes the expected information (-second derivative of the likelihood) with respect to the 3 parameters of the 3PL model. It follows the parametrization $a(θ - b)$.

Arguments

latent::Latent1D : Required. A 1-dimensional Latent1D latent variable.

parameters::Parameters1PL : Required. A 3-parameter logistic parameters object.

Output

A $3 \times 3$ matrix of the expected informations.

Psychometrics.information_latent — Method

information_latent(examinee::AbstractExaminee, items::Vector{<:AbstractItem})

Description

An abstraction of information_latent(latent::AbstractLatent, parameters::AbstractParameters) on an examinee and items. It follows the parametrization $a(θ - b)$.

Arguments

examinee::AbstractExaminee : Required.

items::Vector{<:AbstractItem} : Required.

Psychometrics.information_latent — Method

information_latent(latents_matrix::Matrix{Float64}, parameters_matrix::Matrix{Float64})

Description

It computes the information function (IIF) for item parameters at latents values provided in matrix form. Not suitable for 3PL models, for such a kind of model use informationlatent3PL(). It follows the parametrization $aθ - b$.

Arguments

latents_matrix::Matrix{Float64} : Required. A n_latents x N matrix with latents values.

parameters_matrix::Matrix{Float64} : Required. A (n_latents + 1) x I matrix with item parameters. intercept (b) must be in first row, latents coefficients (aj) in next rows (2, ..., nlatents + 1).

Output

A I x N Float64 matrix.

Psychometrics.information_latent — Method

information_latent(examinees::Vector{<:AbstractExaminee}, items::Vector{<:AbstractItem})

Description

An abstraction of information_latent(latent::AbstractLatent, parameters::AbstractParameters) on examinees and items. It follows the parametrization $a(θ - b)$.

Arguments

examinees::Vector{<:AbstractExaminee} : Required.

items::Vector{<:AbstractItem} : Required.

Psychometrics.information_latent — Method

information_latent(latent::Latent1D, parameters::Parameters1PL)

Description

It computes the information (-second derivative of the likelihood) with respect to the 1-dimensional latent variable under the 1PL model. It follows the parametrization $a(θ - b)$.

Arguments

latent::Latent1D : Required. A 1-dimensional Latent1D latent variable.

parameters::Parameters1PL : Required. A 1-parameter logistic parameters object.

Output

A Float64 scalar.

Psychometrics.information_latent — Method

information_latent(latent::Latent1D, parameters::Parameters2PL)

Description

It computes the information (-second derivative of the likelihood) with respect to the 1-dimensional latent variable under the 2PL model. It follows the parametrization $a(θ - b)$.

Arguments

latent::Latent1D : Required. A 1-dimensional Latent1D latent variable.

parameters::Parameters2PL : Required. A 2-parameter logistic parameters object.

Output

A Float64 scalar.

Psychometrics.information_latent — Method

information_latent(latent::Latent1D, parameters::Parameters3PL)

Description

It computes the information (-second derivative of the likelihood) with respect to the 1-dimensional latent variable under the 3PL model. It follows the parametrization $a(θ - b)$.

Arguments

latent::Latent1D : Required. A 1-dimensional Latent1D latent variable.

parameters::Parameters3PL : Required. A 3-parameter logistic parameters object.

Output

A Float64 scalar.

Psychometrics.information_latent_3PL — Method

information_latent_3PL(latents_matrix::Matrix{Float64}, parameters_matrix::Matrix{Float64})

Description

Only for models which has guessing parameter (c) in last row of parameters_matrix. It computes the information function (IIF) for item parameters at latents values provided in matrix form. It follows the parametrization $aθ - b$.

Arguments

latents_matrix::Matrix{Float64} : Required. A n_latents x N matrix with latents values.

parameters_matrix::Matrix{Float64} : Required. A (n_latents + 1) x I matrix with item parameters. intercept (b) must be in first row, latents coefficients (aj) in next rows (2, ..., nlatents + 1).

Output

A I x N Float64 matrix.

Psychometrics.observed_information_item — Method

observed_information_item(responses::Vector{<:AbstractResponse}, examinees::Vector{<:AbstractExaminee}, items::Vector{<:AbstractItem})

Description

Abstraction to response, item and examinee of observed_information_item(response_val::Float64, latent::Latent1D, parameters::Parameters3PL). It follows the parametrization $a(θ - b)$.

Psychometrics.observed_information_item — Method

observed_information_item(response_val::Float64, latent::Latent1D, parameters::Parameters3PL)

Description

It computes the observed information (-second derivative of the likelihood) with respect to the 3 parameters of the 3PL model. It follows the parametrization $a(θ - b)$.

Arguments

response_val::Float64 : Required. A scalar response.

latent::Latent1D : Required. A 1-dimensional Latent1D latent variable.

parameters::Parameters3PL : Required. A 3-parameter logistic parameters object.

Output

A $3 \times 3$ matrix of the observed informations.