Items

Psychometrics.Item — Type

Item <: AbstractItem

Description

A generic item struct.

Fields

idx::Int64: An integer that identifies the examinee.

id::String: A string that identifies the examinee.

content::Vector{String}: A string vector containing the content features of an item.

parameters::AbstractParameters: A generic item parameters object.

Inner Methods

Item(idx, id, content, parameters) = new(idx, id, content, parameters)

Creates a new 3PL item with custom index, id, content features and item parameters.

Random initilizers

Item(idx, id, content) = new(idx, id, content, Parameters1PL())

Randomly generates a new generic item with custom index, id, content features and default 1PL item parameters (Look at (Parameters1PL)[#Psychometrics.Parameters1PL] for the defaults).

Psychometrics.Item1PL — Type

Item1PL <: AbstractItem

Description

Item struct under the 1-parameter logistic model.

Fields

idx::Int64: An integer that identifies the examinee.

id::String: A string that identifies the examinee.

content::Vector{String}: A string vector containing the content features of an item.

parameters::Parameters1PL: A Parameters1PL object.

Inner Methods

Item1PL(idx, id, content, parameters) = new(idx, id, content, parameters)

Creates a new 1PL item with custom index, id, content features and item parameters.

Random initilizers

Item1PL(idx, id, content) = new(idx, id, content, Parameters1PL())

Randomly generates a new 1PL item with custom index, id, content features and default 1PL item parameters (Look at (Parameters1PL)[#Psychometrics.Parameters1PL] for the defaults).

Psychometrics.Item2PL — Type

Item2PL <: AbstractItem

Description

Item struct under the 2-parameter logistic model.

Fields

idx::Int64: An integer that identifies the examinee.

id::String: A string that identifies the examinee.

content::Vector{String}: A string vector containing the content features of an item.

parameters::Parameters2PL: A Parameters2PL object.

Inner Methods

Item2PL(idx, id, content, parameters) = new(idx, id, content, parameters)

Creates a new 2PL item with custom index, id, content features and item parameters.

Random initilizers

Item2PL(idx, id, content) = new(idx, id, content, Parameters2PL())

Randomly generates a new 2PL item with custom index, id, content features and default 2PL item parameters (Look at (Parameters2PL)[#Psychometrics.Parameters2PL] for the defaults).

Psychometrics.Item3PL — Type

Item3PL <: AbstractItem

Description

Item struct under the 3-parameter logistic model.

Fields

idx::Int64: An integer that identifies the examinee.

id::String: A string that identifies the examinee.

content::Vector{String}: A string vector containing the content features of an item.

parameters::Parameters2PL: A Parameters3PL object.

Inner Methods

Item3PL(idx, id, content, parameters) = new(idx, id, content, parameters)

Creates a new 3PL item with custom index, id, content features and item parameters.

Random initilizers

Item3PL(idx, id, content) = new(idx, id, content, Parameters3PL())

Randomly generates a new 3PL item with custom index, id, content features and default 3PL item parameters (Look at (Parameters3PL)[#Psychometrics.Parameters3PL] for the defaults).

Psychometrics.get_item_by_idx — Method

get_item_by_idx(item_idx::Int64, items::Vector{<:AbstractItem})

It returns the item with index item_idx from a vector of <:AbstractItem.