Latent Variables

Psychometrics.Latent — Type

Latent <: AbstractLatent

Description

Generic latent variable struct.

Fields

- **`val::Vector{Float64}`**
- **`bounds::Vector{Vector{Float64}}`**
- **`prior::Union{Distributions.ContinuousUnivariateDistribution, Distributions.MultivariateDistribution}`**
- **`posterior::Union{Distributions.ContinuousUnivariateDistribution, Distributions.MultivariateDistribution}`**
- **`chain::Vector{Vector{Float64}} `**
- **`expected_information::Matrix{Float64}`**

Factories

Latent(names, val, bounds, prior, posterior, chain, expected_information) = new(val, bounds, prior, posterior, chain, expected_information)

Creates a new generic latent variable with custom fields.

Random Initializers

Latent(bounds, posterior)

Randomly generates a value for the generic latent variable and assigns a custom univariate or multivariate distribution to prior and posterior with specific bounds.