Public Documentation

Documentation for Psychometrics.jl's public interface.

See the Internals Documentation Section of the manual for internal package docs covering unexported functions.

Index

Public Documentation

Index

Structs and Abstracts

Exported Methods

Structs and Abstracts

Psychometrics.AbstractLatent — Type

AbstractLatent

Psychometrics.AbstractParameters — Type

Description

An abstract type which, at the moment, has only the abstract AbstractParametersBinary as subtype. Thus, it works only with binary (dichotomous) responses. The available mutables for the latter are Parameters1PL, Parameters2PL, Parameters3PL. They contain the details about the item parameters under the 1-parameter logistic (1PL), 2-parameters logistic (2PL) and 3-parameter logistic (3PL) IRT models, respectively.

Example

The object Parameters2PL has the fields difficulty b::Float64 and discrimination a::Float64. It is possible to define the Bayesian priors and posterior by assigning multivariate distributions from the package Distributions to the fields prior::Distributions.MultivariateDistribution and posterior::Distributions.MultivariateDistribution.

Psychometrics.AbstractResponse — Type

AbstractResponse

Psychometrics.Examinee — Type

Description

An immutable containing the information about the examinee (the test taker).

Fields

idx::Int64: An integer that identifies the Examinee in this session.

id::String: A string that identifies the Examinee.

latent::Latent: A mutable latent struct associated with the Examinee.

Factories

Examinee(idx, id, latent) = new(idx, id, latent)

Creates a new Examinee with custom index, id and a generic latent variable.

Random initializers

Examinee(idx, id) = new(idx, id, Latent1D())

Randomly generates an Examinee with custom index and id and with a default 1-dimensional latent variable (Look at (Latent1D)[#Psychometrics.Latent1D] for the defaults).

Psychometrics.Item — Type

Item <: AbstractItem

Description

An immutable containing information about an item (a question in a test), e.g. id::String the item identifier, calibrated::Bool says if it is a field item (false) or an operational item (true), and the field parameters::AbstractParameter which accepts a mutable item parameter object

Fields

idx::Int64: An integer that identifies the item in this session.

id::String: A string that identifies the examinee.

content::Vector{String}: A string vector containing the content features of an item.

parameters::AbstractParameters: A generic item parameters object.

calibrated::Bool: Tells if the item has been already calibrated.

Factories

Item(idx, id, content, parameters) = new(idx, id, content, parameters)

Creates a new generic item with custom index, id, content features and item parameters.

Random initilizers

Item(idx, id) = new(idx, id, "", Parameters1PL(), true)
Item(idx, id, content) = new(idx, id, content, Parameters1PL(), true)

Randomly generates a new generic calibrated item with custom index, id, content features and default 1PL item parameters (Look at (Parameters1PL)[#Psychometrics.Parameters1PL] for the defaults).

Psychometrics.Latent1D — Type

Latent1D <: AbstractLatent

Description

A mutable describing an univariate latent variable. The field val::Float64 holds the estimate of the ability of the examinee, like the item parameters, also the latent variables can have a prior or a posterior, assigning them to the fields prior and posterior, respectively.

Fields

val::Float64

bounds::Vector{Float64}

prior::Distributions.ContinuousUnivariateDistribution

posterior::Distributions.ContinuousUnivariateDistribution

chain::Vector{Float64}

expected_information::Float64

Factories

Latent1D(val, bounds, prior, posterior, chain, expected_information) = new(val, bounds, prior, posterior, chain, expected_information)

Creates a new 1-dimensional latent variable with custom fields.

Random Initializers

Latent1D()

Randomly generates a value for the 1-dimensional latent variable and assigns a default standardized Gaussian prior and posterior Latent1D(dist, bounds)

Same as Latent1D(dist, bounds) but with value equal to val Latent1D(val)

Randomly generates a value for the 1-dimensional latent variable and assigns a custom univariate distribution to prior and posterior with specific bounds.

Psychometrics.LatentND — Type

LatentND <: AbstractLatent

Description

A mutable describing a multivariate latent variable.

Fields

names::Vector{String}

val::Vector{Float64}

bounds::Vector{Vector{Float64}}

prior::Distributions.MultivariateDistribution

posterior::Distributions.MultivariateDistribution

chain::Vector{Vector{Float64}}

expected_information::Matrix{Float64}

Factories

LatentND(names, val, bounds, prior, posterior, chain, expectedinformation) = new(names, val, bounds, prior, posterior, chain, expectedinformation) Creates a new N-dimensional latent variable with custom fields.

Random Initializers

LatentND()

Randomly generates a value for the N-dimensional latent variable and assigns a default standardized Gaussian prior and posterior LatentND(dist, bounds)

Randomly generates a value for the N-dimensional latent variable and assigns a custom multivariate distribution to prior and posterior with specific bounds.

Psychometrics.Parameters1PL — Type

Parameters1PL <: AbstractParametersBinary

Contains information of a set of item parameters (values, bounds, priors, posteiors, chains, expected Fisher information, calibrated) under the 1-parameter logistic model.

Psychometrics.Parameters2PL — Type

Parameters2PL <: AbstractParametersBinary

Contains information of a set of item parameters (values, bounds, priors, posteiors, chains, expected Fisher information, calibrated) under the 2-parameter logistic model.

Psychometrics.Parameters3PL — Type

Contains information of a set of item parameters (values, bounds, priors, posteiors, chains, expected Fisher information, calibrated) under the 3-parameter logistic model.

Psychometrics.ParametersNPL — Type

ParametersNPL <: AbstractParametersBinary

Contains info about the difficulty of an item under the N-parameter logistic model.

Psychometrics.Response — Type

Response <: AbstractResponse

Description

An immutable which holds the information about response, such as the identifier of the examinee who gave the response, examinee_id::String, the identifier of the answered item item_idx::String, and the answer starting and ending times start_time::Dates.DateTime and end_time::Dates.DateTime.

Each of the mentioned structs has a random default factory, callable by using the name of the struct followed by ().

Example: Parameters2PL() generates a set of 2PL item parameters from the product distribution (independent bivariate).

Distributions.Product([
   Distributions.LogNormal(0, 0.25),
   Distributions.Normal(0, 1)
])

The immutable structs Item and Examinee need the identification fields id and idx to be randomly generated.

Example: `Item(1, \"item1\")generates an operational item with an emptycontent` description, and a default 1PL item parameter._

Exported Methods

Items

Examinees

Responses

Probability (Item Characteristic Function)

Information (Item Information Function)

Likelihood