Log Likelihood

Psychometrics.log_likelihood — Method

log_likelihood(responses::Vector{<:AbstractResponse}, g_item::Vector{Float64}, g_latent::Vector{Float64})

It computes the log likelihood for a vector of responses responses. It updates also the gradient vectors.

Psychometrics.log_likelihood — Method

log_likelihood(response::AbstractResponse, g_item::Vector{Float64}, g_latent::Vector{Float64})

It computes the log likelihood for a response response. It updates also the gradient vectors.

Psychometrics.log_likelihood — Method

log_likelihood(response::AbstractResponse, g_item::Vector{Float64}, g_latent::Vector{Float64})

It computes the log likelihood for a response response.

Psychometrics.log_likelihood — Method

log_likelihood(latent::Latent1D, parameters::AbstractParameters, response_val::Float64, g_item::Vector{Float64}, g_latent::Vector{Float64})

It computes the log likelihood for a 1-dimensional latent variable and item parameters parameters with answer response_val. It updates also the gradient vectors.

Psychometrics.log_likelihood — Method

log_likelihood(latent::Latent1D, parameters::AbstractParameters, response_val::Float64)

It computes the log likelihood for a 1-dimensional latent variable and item parameters parameters with answer response_val.