Examinees

Struct

Psychometrics.Examinee — Type

Examinee <: AbstractExaminee

Description

An immutable containing the information about the examinee (the test taker).

Fields

idx::Int64: An integer that identifies the Examinee in this session.

id::String: A string that identifies the Examinee.

latent::Latent: A mutable latent struct associated with the Examinee.

Factories

Examinee(idx::Int64, id::String, latent::AbstractLatent) = new(idx, id, latent)
Examinee(idx::Int64, latent::AbstractLatent) = new(idx, string(idx), latent)

Creates a new Examinee with custom index, (id) and a latent variable.

Random default initializers (1-D latent)

Examinee(idx::Int64) = new(idx, string(idx), Latent1D())
Examinee(idx::Int64, id::String) = new(idx, id, Latent1D())

Description

Randomly generates an Examinee with custom index and id and with a default 1-dimensional latent variable (Look at (Latent1D)[#Psychometrics.Latent1D] for the defaults).

Methods for a Single Examinee

Psychometrics.get_examinee_by_id — Method

get_examinee_by_id(examinee_id::String, examinees::Vector{<:AbstractExaminee})

It returns the Examinee with index examinee_id from a Vector of AbstractExaminee.

Psychometrics.empty_chain! — Method

empty_chain!(examinee::AbstractExaminee)

Psychometrics.set_val! — Method

set_val!(examinee::AbstractExaminee, val::Float64)

Psychometrics.set_val_from_chain! — Method

set_val_from_chain!(examinee::AbstractExaminee)

Psychometrics.update_estimate! — Method

update_estimate!(examinee::AbstractExaminee; sampling = true)

Psychometrics.chain_append! — Method

chain_append!(examinee::AbstractExaminee; sampling = false)

Psychometrics.set_prior! — Method

set_prior!(
    examinee::AbstractExaminee,
    prior::Union{Distributions.DiscreteUnivariateDistribution, Distributions.ContinuousUnivariateDistribution}
)

Missing docstring.

Missing docstring for get_latents(examinee::AbstractExaminee). Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_latents_vals(examinee::AbstractExaminee). Check Documenter's build log for details.

Methods for a Vector of Examinees

Psychometrics.get_latents — Method

get_latents(examinees::Vector{<:AbstractExaminee})

#Description

Returns the latent objects of a vector of examinees as a vector of subtypes of AbstractLatent.

Psychometrics.get_latents_vals — Method

get_latents_vals(examinees::Vector{<:AbstractExaminee})

#Description

Returns a matrix with latent values displayed by row.

Psychometrics.set_prior! — Method

set_prior!(
    examinees::Vector{<:AbstractExaminee},
    prior::Union{Distributions.DiscreteUnivariateDistribution, Distributions.ContinuousUnivariateDistribution}
)