Information (Item Information Function)

Information wrt $\theta$

Values in Matrix form

Psychometrics.information_latent — Method

information_latent(latents_matrix::Matrix{Float64}, parameters_matrix::Matrix{Float64})

Description

It computes the information function (IIF) for item parameters at latents values provided in matrix form. Not suitable for 3PL models, for such a kind of model use information_latent_3PL(). It follows the parametrization $a \theta - b$. See the docs of information_latent(examinee::AbstractExaminee, item::AbstractItem) for details.

Arguments

latents_matrix::Matrix{Float64} : Required. A $n_latents \times N$ matrix with latents values.

parameters_matrix::Matrix{Float64} : Required. A $(n_latents + 1) \times I$ matrix with item parameters. intercept (b) must be in first row, latents coefficients $(a_j)$ in next rows $(2, \ldots, n_latents + 1)$.

Output

A I x N Float64 matrix.

Psychometrics.information_latent_3PL — Method

informationlatent3PL( latentsmatrix::Matrix{Float64}, parametersmatrix::Matrix{Float64}, )

Description

Only for models which has guessing parameter (c) in last row of parameters_matrix. It computes the information function (IIF) for item parameters at latents values provided in matrix form. It follows the parametrization $a \theta - b$.

Arguments

latents_matrix::Matrix{Float64} : Required. A n_latents x N matrix with latents values.

parameters_matrix::Matrix{Float64} : Required. A (n_latents + 1) x I matrix with item parameters. intercept (b) must be in first row, latents coefficients (aj) in next rows (2, ..., nlatents + 1).

Output

A I x N Float64 matrix.

Using Structs

Psychometrics.information_latent — Method

information_latent(examinee::AbstractExaminee, item::AbstractItem)

Description

An abstraction of _information_latent(latent::AbstractLatent, parameters::AbstractParametersBinary) on an examinee and an item. It follows the parametrization $a(θ - b)$.

Arguments

examinee::AbstractExaminee : Required.

item::AbstractItem : Required.

Psychometrics.information_latent — Method

information_latent(examinee::AbstractExaminee, items::Vector{<:AbstractItem})

Description

An abstraction of _information_latent(latent::AbstractLatent, parameters::AbstractParametersBinary) on an examinee and items. It follows the parametrization $a(θ - b)$.

Arguments

examinee::AbstractExaminee : Required.

items::Vector{<:AbstractItem} : Required.

Example

Compute the Fisher information for the latents of the examinees (second derivatives of the likelihood with respect to the latent $\theta$) and each item.

$E_\theta_n \[ I(\theta_n | b_i) \]$

examinees = [Examinee() for n = 1 : 10]; #default examinee factory
items = [Item() for i = 1 : 30]; #default item factory
probability(examinees, items) #compute the probability
information_latent(examinees, items) #compute the information wrt θ   

Psychometrics.information_latent — Method

information_latent( examinees::Vector{<:AbstractExaminee}, items::Vector{<:AbstractItem}, )

Description

An abstraction of _information_latent(latent::AbstractLatent, parameters::AbstractParametersBinary) on examinees and items. It follows the parametrization $a(θ - b)$.

Arguments

examinees::Vector{<:AbstractExaminee} : Required.

items::Vector{<:AbstractItem} : Required.

Information wrt Item Parameters

Psychometrics.expected_information_item — Method

expectedinformationitem( items::Vector{<:AbstractItem}, examinees::Vector{<:AbstractExaminee}, )

Description

Abstraction of expectedinformation_item(latent, parameters) on Vector{<:AbstractExaminee} and items::Vector{<:AbstractItem}. It follows the parametrization $a(θ - b)$.

Arguments

items::Vector{<:AbstractItem} : Required. Size I x 1.

examinees::Vector{<:AbstractExaminee} : Required. Size N x 1.

Output

A $N x I$ matrix of the expected informations matrices.

Psychometrics.observed_information_item — Method

observedinformationitem( items::Vector{<:AbstractItem}, examinees::Vector{<:AbstractExaminee}, responses::Vector{<:AbstractResponse}, )

Description

Abstraction to response, item and examinee of _observed_information_item(response_val::Float64, latent::Latent1D, parameters::Parameters3PL). It follows the parametrization $a(θ - b)$.