Likelihood

Psychometrics.log_likelihood — Method

log_likelihood(
    response_val::Float64,
    latent_val::Float64,
    parameters::AbstractParametersBinary,
)

Description

It computes the log likelihood for a latent value and item parameters parameters with answer response_val.

Psychometrics.log_likelihood — Method

log_likelihood(
    response::AbstractResponse,
    examinee::AbstractExaminee,
    item::AbstractItem,
)

Description

It computes the log likelihood for a response.

Psychometrics.log_likelihood — Method

log_likelihood(
    response::AbstractResponse,
    examinee::AbstractExaminee,
    item::AbstractItem,
    g_item::Vector{Float64},
    g_latent::Vector{Float64},
)

Description

It computes the log likelihood for a response response. It updates also the gradient vectors.

Psychometrics.log_likelihood — Method

log_likelihood(
    responses::Vector{<:AbstractResponse},
    examinees::Vector{<:AbstractExaminee},
    items::Vector{<:AbstractItem},
)

Description

It computes the log likelihood for a vector of responses responses.

Psychometrics.log_likelihood — Method

log_likelihood(
    responses::Vector{<:AbstractResponse},
    examinees::Vector{<:AbstractExaminee},
    items::Vector{<:AbstractItem},
    g_item::Vector{Float64},
    g_latent::Vector{Float64},
)

Description

It computes the log likelihood for a vector of responses responses. It updates also the gradient vectors.

Psychometrics.likelihood — Method

likelihood(
    response::AbstractResponse,
    examinee::AbstractExaminee,
    item::AbstractItem,
)

Description

It computes the log likelihood for a response.

Psychometrics.likelihood — Method

likelihood(
    responses::Vector{<:AbstractResponse},
    examinees::Vector{<:AbstractExaminee},
    items::Vector{<:AbstractItem},
)

Description

It computes the log likelihood for a vector of responses responses.