Online Calibration Utilities

Psychometrics.find_best_examinee — Method

find_best_examinee(item, examinees; method = "D")

Description

Finds the best examinee among the examinees vector, for the item using maximum expected information criterion.

Arguments

item::AbstractItem: Required. Set of items in which searching for the best.

examinees::Vector{<:AbstractExaminee}: Required. The examinee.

method: Optional. "D" | "A". D stands for D-optimality (determinant of the expected information matrix), A stands for A-optimality (trace of the expected information matrix).

Output

It returns the idx of the best examinee.

Psychometrics.find_best_item — Method

find_best_item(examinee, items)

Description

Finds the best item for the examinee among the vector items using maximum information criterion.

Arguments

examinee::AbstractExaminee: The examinee.

items:::Vector{<:AbstractItem}: Set of items in which searching for the best.

method: Optional. "max-info" | "D-gain".

max-info takes the item with the maximum latent information, D-gain takes the item with the maximum gain in the determinant of the expected information matrix if examinee would answer to the item (D-VC in Ren, van der Linden, Diao, 2017).

Output

It returns an item of generic type.

References

[RenDiao2017]

Psychometrics.A_gain_method — Method

A_gain_method(item::AbstractItem, examinee::AbstractExaminee)

Description

Computes the gain in the trace of the expected information matrix for an item.

Arguments

item::AbstractItem: The item.

examinee::AbstractExaminee: The examinee at which computing the information.

Output

It returns a Float64 scalar.

Psychometrics._A_gain_method — Method

_A_gain_method(parameters::Parameters1PL, latent::Latent1D)

Psychometrics._A_gain_method — Method

_A_gain_method(parameters::Parameters2PL, latent::Latent1D)

Psychometrics.A_inv_gain_method — Method

A_inv_gain_method(item::AbstractItem, examinee::AbstractExaminee)

Description

Computes the gain in the trace of the inverse of the expected information matrix for an item.

Arguments

item::AbstractItem: The item.

examinee::AbstractExaminee: The examinee at which computing the information.

Output

It returns a Float64 scalar.

Psychometrics._A_inv_gain_method — Method

_A_inv_gain_method(parameters::Parameters1PL, latent::Latent1D)

Psychometrics._A_inv_gain_method — Method

_A_inv_gain_method(parameters::Parameters2PL, latent::Latent1D)

Psychometrics.A_method — Method

A_method(item::AbstractItem, examinee::AbstractExaminee)

Description

Computes the trace of the expected information matrix for an item and an examinee.

Arguments

item::AbstractItem: The item.

examinee::AbstractExaminee: The examinee at which computing the information.

Output

It returns a Float64 scalar.

Psychometrics._A_method — Method

_A_method(parameters::Parameters1PL, latent::Latent1D)

Psychometrics._A_method — Method

_A_method(parameters::Parameters2PL, latent::Latent1D)

Psychometrics.D_gain_method — Method

D_gain_method(item::AbstractItem, examinee::AbstractExaminee)

Description

Computes the gain in the determinant of the expected information matrix for an item.

Arguments

item::AbstractItem: The item.

examinee::AbstractExaminee: The examinee at which computing the information.

Output

It returns a Float64 scalar.

References

Ren H, van der Linden WJ, Diao Q. Continuous online item calibration: Parameter recovery and item calibration. Psychometrika. 2017;82:498–522. doi: 10.1007/s11336-017-9553-1.

Psychometrics._D_gain_method — Method

_D_gain_method(parameters::Parameters1PL, latent::Latent1D)

Psychometrics._D_gain_method — Method

_D_gain_method(parameters::Parameters2PL, latent::Latent1D)

Psychometrics.D_inv_gain_method — Method

D_inv_gain_method(item::AbstractItem, examinee::AbstractExaminee)

Description

It computes the gain in the inverse of the expected information of the item.

Arguments

item::AbstractItem: The item.

examinee::AbstractExaminee: The examinee at which computing the information.

Output

It returns a Float64 scalar.

References

Yinhong He & Ping Chen, 2020. "Optimal Online Calibration Designs for Item Replenishment in Adaptive Testing," Psychometrika, Springer;The Psychometric Society, vol. 85(1), pages 35-55, March.

Psychometrics._D_inv_gain_method — Method

_D_inv_gain_method(parameters::Parameters1PL, latent::Latent1D)

Psychometrics._D_inv_gain_method — Method

_D_inv_gain_method(parameters::Parameters2PL, latent::Latent1D)

RenDiao2017Ren H, van der Linden WJ, Diao Q. Continuous online item calibration: Parameter recovery and item calibration. Psychometrika. 2017;82:498–522. doi: 10.1007/s11336-017-9553-1.