Responses

Psychometrics.get_examinees_by_item_id — Method

get_examinees_by_item_id(
    item_id::String,
    responses::Vector{<:AbstractResponse},
    examinees::Vector{<:AbstractExaminee},
)

Description

It returns the examinees who answered to the item with id item_id.

Psychometrics.get_items_by_examinee_id — Method

get_items_by_examinee_id(
    examinee_id::String,
    responses::Vector{<:AbstractResponse},
    items::Vector{<:AbstractItem},
)

Description

It returns the items answered by the examinee with id examinee_id.

Psychometrics.add_response! — Method

add_response!(response::AbstractResponse, responses::Vector{<:AbstractResponse})

Description

Push the response in the response vector responses.

Psychometrics.get_responses_by_examinee_id — Method

get_responses_by_examinee_id(examinee_id::String, responses::Vector{<:AbstractResponse})

Description

It returns the vector of responses given by examinee with id = id.

Psychometrics.get_responses_by_item_id — Method

get_responses_by_item_id(item_id::String, responses::Vector{<:AbstractResponse})

Description

It returns the vector of responses to item with id equal to item_id.

Psychometrics.get_responses_by_item_idx — Method

get_responses_by_item_idx(item_idx::Int64, responses::Vector{<:AbstractResponse}; sorted = true)

Description

It returns the vector of responses to item with idx equal to item_idx. The vector of responses is sorted by examinee_idx if sorted = true.

Psychometrics.answer — Method

answer(examinee::AbstractExaminee, item::AbstractItem)

Description

Randomly generate a response by examinee to a dichotomous (binary) item.

Psychometrics.answer — Method

answer(examinee::AbstractExaminee, items::Vector{<:AbstractItem})

Description

Randomly generate a response by examinee to dichotomous (binary) items.

Psychometrics.answer — Method

answer(examinee_id::String, item_id::String, examinees::Vector{<:AbstractExaminee}, items::Vector{<:AbstractItem})

Description

Randomly generate a response by Examinee with index examinee_id to a dichotomous (binary) item with index item_id.

Psychometrics.answer — Method

answer(examinees::Vector{<:AbstractExaminee}, items::Vector{<:AbstractItem})

Description

Randomly generate responses by all the examinees in examinees to items in items.

Psychometrics.get_design_matrix — Method

get_design_matrix(responses::Vector{Response}, I::Int64, N::Int64)

Description

Returns the $I imes N$ design matrix.

Psychometrics.get_response_matrix — Method

get_response_matrix(responses::Vector{Response}, I::Int64, N::Int64)

Description

Transform vector of Responses in a $I imes N$ response matrix. A non given answer has value 0.0.

Psychometrics.get_responses — Method

get_responses(response_matrix::Matrix{Float64}, design_matrix::Matrix{Float64}, items::Vector{<:AbstractItem}, examinees::Vector{<:AbstractExaminee})

Transforms a $I imes N$ response matrix in a vector of Responses given a valid design_matrix, a vector of Items and a vector of Examinees.

Psychometrics.get_items_idx_answered_by_examinee — Method

get_items_idx_answered_by_examinee(
    examinee::AbstractExaminee,
    responses::Vector{<:AbstractResponse},
)

Description

Returns the idx of the items answered by examinee.

Psychometrics.get_examinees_idx_who_answered_item — Method

get_examinees_idx_who_answered_item(
    item::AbstractItem,
    responses::Vector{<:AbstractResponse},
)

Description

Returns the idx of the examinees who answered to item.