Probability (Item Characteristic Function) (Internals)

1PL

Psychometrics.__probability — Method

__probability(latent_val::Float64, parameters::Parameters1PL)

Description

It computes the probability (ICF) of a correct response for item parameters under the 1PL model at latent_val point. It follows the parametrization $a(θ - b)$.

Arguments

latent_val::Float64 : Required. The point in the latent space in which compute the probability.

parameters::Parameters1PL : Required. A 1-parameter logistic parameters object.

Output

A Float64 scalar.

Psychometrics._probability — Method

_probability(latent::Latent1D, parameters::Parameters1PL)

Description

It computes the probability (ICF) of a correct response for item parameters under the 1PL model at Latent1D point. It follows the parametrization $a(θ - b)$.

Arguments

latent::Latent1D : Required. A 1-dimensional Latent1D latent variable.

parameters::Parameters1PL : Required. A 1-parameter logistic parameters object.

Output

A Float64 scalar.

Psychometrics._probability — Method

_probability(latent::Latent1D, parameters::Parameters1PL, g_item::Vector{Float64}, g_latent::Vector{Float64})

Description

It computes the probability (ICF) of a correct response for item parameters under the 1PL model at Latent1D point. It updates the gradient vectors if they are not empty. It follows the parametrization $a(θ - b)$.

Arguments

latent::Latent1D : Required. A 1-dimensional Latent1D latent variable.

parameters::Parameters1PL : Required. A 1-parameter logistic parameters object.

g_item::Vector{Float64} : Optional. Vector of gradients with respect to item parameters.

g_latent::Vector{Float64} : Optional. Vector of gradients with respect to the 1-dimensional latent.

Output

A Float64 scalar. Updates gradient vectors.

2PL

Psychometrics.__probability — Method

__probability(latent_val::Float64, parameters::Parameters2PL)

Description

It computes the probability (ICF) of a correct response for item parameters under the 2PL model at latent_val point. It follows the parametrization $a(θ - b)$.

Arguments

latent_val::Float64 : Required. The point in the latent space in which compute the probability.

parameters::Parameters2PL : Required. A 2-parameter logistic parameters object.

Output

A Float64 scalar.

Psychometrics._probability — Method

_probability(latent::Latent1D, parameters::Parameters2PL)

Description

It computes the probability (ICF) of a correct response for item parameters under the 2PL model at Latent1D point. It follows the parametrization $a(θ - b)$.

Arguments

latent::Latent1D : Required. A 1-dimensional Latent1D latent variable.

parameters::Parameters2PL : Required. A 2-parameter logistic parameters object.

Output

A Float64 scalar.

Psychometrics._probability — Method

_probability(latent::Latent1D, parameters::Parameters2PL, g_item::Vector{Float64}, g_latent::Vector{Float64})

Description

It computes the probability (ICF) of a correct response for item parameters under the 2PL model at Latent1D point. It updates the gradient vectors if they are not empty. It follows the parametrization $a(θ - b)$.

Arguments

latent::Latent1D : Required. A 1-dimensional Latent1D latent variable.

parameters::Parameters2PL : Required. A 2-parameter logistic parameters object.

g_item::Vector{Float64} : Optional. Vector of gradients with respect to item parameters.

g_latent::Vector{Float64} : Optional. Vector of gradients with respect to the 1-dimensional latent.

Output

A Float64 scalar. Updates the gradient vectors.

Example

Probability of a correct response is computed for each examinee and each item.

3PL

__probability(latent_val::Float64, parameters::Parameters3PL)
_probability(latent::Latent1D, parameters::Parameters3PL)
_probability(latent::Latent1D, parameters::Parameters3PL,  g_item::Vector{Float64}, g_latent::Vector{Float64})
__probability(latent_vals::Vector{Float64}, parameters::Parameters3PL)
_probability(latent::Latent1D, parameters::Parameters3PL,  g_item::Vector{Float64}, g_latent::Vector{Float64})

```