Probability (Item Characteristic Function)

Psychometrics.__probability — Method

__probability(latent_val::Float64, parameters::Parameters3PL)

Description

It computes the probability (ICF) of a correct response for item parameters under the 3PL model at latent_val point. It follows the parametrization $a(θ - b)$.

Arguments

latent_val::Float64 : Required. The point in the latent space in which compute the probability.

parameters::Parameters3PL : Required. A 3-parameter logistic parameters object.

Output

A Float64 scalar.

Psychometrics.__probability — Method

__probability(latent_vals::Vector{Float64}, parameters::ParametersNPL)

Description

It computes the probability (ICF) of a correct response for item parameters under the 3PL model at latent_vals points. N-dimensional latent. It follows the parametrization $a(θ - b)$.

Arguments

latent_vals::Vector{Float64} : Required. The points in the latent space in which compute the probability.

parameters::ParametersNPL : Required. A 2-parameter logistic parameters object suitbale for N-dimensional latent.

Output

A Float64 scalar.

Psychometrics._probability — Method

_probability(latent::Latent1D, parameters::Parameters3PL,  g_item::Vector{Float64}, g_latent::Vector{Float64})

Description

It computes the probability (ICF) of a correct response for item parameters under the 3PL model at Latent1D point. It updates the gradient vectors if they are not empty. It follows the parametrization $a(θ - b)$.

Arguments

latent::Latent1D : Required. A 1-dimensional Latent1D latent variable.

parameters::Parameters3PL : Required. A 3-parameter logistic parameters object.

g_item::Vector{Float64} : Optional. Vector of gradients with respect to item parameters.

g_latent::Vector{Float64} : Optional. Vector of gradients with respect to the 1-dimensional latent.

Output

A Float64 scalar. Updates the gradient vectors.

Psychometrics._probability — Method

_probability(latent::Latent1D, parameters::Parameters3PL)

Description

It computes the probability (ICF) of a correct response for item parameters under the 3PL model at Latent1D point. It follows the parametrization $a(θ - b)$.

Arguments

latent::Latent1D : Required. A 1-dimensional Latent1D latent variable.

parameters::Parameters3PL : Required. A 3-parameter logistic parameters object.

Output

A Float64 scalar.

Psychometrics._probability — Method

_probability(latent::LatentND, parameters::ParametersNPL)

Description

It computes the probability (ICF) of a correct response for item parameters under the 3PL model at Latent1D point. It updates the gradient vectors if they are not empty. It follows the parametrization $a(θ - b)$.

Arguments

latent::LatentND : Required. A N-dimensional latent variable.

parameters::ParametersNPL : Required. A 2-parameter logistic parameters object suitable for a N-dimensional latent.

Output

A Float64 scalar.

Psychometrics.probability — Method

probability(examinees::Vector{<:AbstractExaminee}, items::Vector{<:AbstractItem})

Description

It computes the probability (ICF) that a vector of examinees answers correctly at a vector of items.

Arguments

examinees::Vector{<:AbstractExaminee} : Required. A vector of Examinees.

items::Vector{<:AbstractItem} : Required. A vector of Items.

Output

A I x N Float64 matrix.

Psychometrics.probability — Method

probability(items::Vector{<:AbstractItem}, examinees::Vector{<:AbstractExaminee})

Description

It computes the probability (ICF) that a vector of examinees answers correctly at a vector of items.

Arguments

examinees::Vector{<:AbstractExaminee} : Required. A vector of Examinees.

items::Vector{<:AbstractItem} : Required. A vector of Items.

Output

A I x N Float64 matrix.