Examinee Structs and Methods

Psychometrics.answer — Method

answer(examinees::Vector{<:AbstractExaminee}, items::Vector{<:AbstractItem})

Randomly generate responses by all the examinees in examinees to items in items.

Psychometrics.answer — Method

answer(examinee_idx::Int64, item_idx::Int64, examinees::Vector{<:AbstractExaminee}, items::Vector{<:AbstractItem})

Randomly generate a response by examinee with index examinee_idx to item with index item_idx.

Psychometrics.answer — Method

answer(examinee::AbstractExaminee, item::AbstractParameters)

Randomly generate a response by examinee to item.

Psychometrics.get_examinee_by_idx — Method

get_examinee_by_idx!(examinee_idx::Int64, examinees::Vector{<:AbstractExaminee})

It returns the examinee with index examinee_idx from a vector of ::AbstractExaminee.