Examinees

Psychometrics.Examinee — Type

Examinee <: AbstractExaminee

Description

Examinee struct with a generic latent variable.

Fields

id::String: A string that identifies the Examinee.

latent::Latent: A generic latent variable associated with the Examinee.

Factories

Examinee1D(id, latent) = new(id, latent)

Creates a new Examinee with custom index, id and a generic latent variable.

Random initializers

Examinee(id) = Examinee1D(id)

Randomly generates an Examinee with custom index and id and with a default 1-dimensional latent variable (Look at (Latent1D)[#Psychometrics.Latent1D] for the defaults).

Psychometrics.Examinee1D — Type

Examinee1D <: AbstractExaminee

Description

Examinee struct with a 1-dimensional latent variable.

Fields

id::String: A string that identifies the Examinee.

latent::Latent1D: A 1-dimensional latent variable associated with the Examinee.

Factories

Examinee1D(id, latent) = new(id, latent)

Creates a new Examinee with custom index, id and 1-dimensional latent variable.

Random initializers

Examinee1D(id) = new(id, Latent1D())

Randomly generates an Examinee with custom index and id and with a default 1-dimensional latent variable (Look at (Latent1D)[#Psychometrics.Latent1D] for the defaults).

Psychometrics.answer — Method

answer(Examinee::AbstractExaminee, item::AbstractParameters)

Randomly generate a response by Examinee to item.

Psychometrics.answer — Method

answer(examinees:AbstractExaminee}, items::Vector{<:AbstractItem})

Randomly generate responses by all the examinees in examinees to items in items.

Psychometrics.answer — Method

answer(examinee_idx::Int64, item_idx::Int64, examinees::Vector{<:AbstractExaminee}, items::Vector{<:AbstractItem})

Randomly generate a response by Examinee with index examinee_idx to item with index item_idx.

Psychometrics.get_examinee_by_idx — Method

get_examinee_by_idx!(examinee_idx::Int64, examinees::Dict{Int64,<:AbstractExaminee})

It returns the Examinee with index examinee_idx from a Dict of AbstractExaminee.