Examinees

Psychometrics.Examinee — Type

Examinee <: AbstractExaminee

Description

An immutable containing the information about the examinee (the test taker).

Fields

idx::Int64: An integer that identifies the Examinee in this session.

id::String: A string that identifies the Examinee.

latent::Latent: A mutable latent struct associated with the Examinee.

Factories

Examinee(idx::Int64, id::String, latent::AbstractLatent) = new(idx, id, latent)
Examinee(idx::Int64, latent::AbstractLatent) = new(idx, string(idx), latent)

Creates a new Examinee with custom index, (id) and a latent variable.

Random default initializers (1-D latent)

Examinee(idx::Int64) = new(idx, string(idx), Latent1D())
Examinee(idx::Int64, id::String) = new(idx, id, Latent1D())

Description

Randomly generates an Examinee with custom index and id and with a default 1-dimensional latent variable (Look at (Latent1D)[#Psychometrics.Latent1D] for the defaults).

Psychometrics.chain_append! — Method

chain_append!(examinee::AbstractExaminee; sampling = false)

Psychometrics.empty_chain! — Method

empty_chain!(examinee::AbstractExaminee)

Psychometrics.get_examinee_by_id — Method

get_examinee_by_id(examinee_id::String, examinees::Vector{<:AbstractExaminee})

It returns the Examinee with index examinee_id from a Vector of AbstractExaminee.

Psychometrics.set_prior! — Method

set_prior!(
    examinee::AbstractExaminee,
    prior::Union{Distributions.DiscreteUnivariateDistribution, Distributions.ContinuousUnivariateDistribution}
)

Psychometrics.set_val! — Method

set_val!(examinee::AbstractExaminee, val::Float64)

Psychometrics.set_val_from_chain! — Method

set_val_from_chain!(examinee::AbstractExaminee)

Psychometrics.update_estimate! — Method

update_estimate!(examinee::AbstractExaminee; sampling = true)

Psychometrics.AbstractLatent — Type

AbstractLatent

Psychometrics.Latent1D — Type

Latent1D <: AbstractLatent

Description

A mutable describing an univariate latent variable. The field val::Float64 holds the estimate of the ability of the examinee, like the item parameters, also the latent variables can have a prior or a posterior, assigning them to the fields prior and posterior, respectively.

Fields

val::Float64

bounds::Vector{Float64}

prior::Union{Distributions.ContinuousUnivariateDistribution, Distributions.DiscreteUnivariateDistribution}

posterior::Union{Distributions.ContinuousUnivariateDistribution, Distributions.DiscreteUnivariateDistribution}

likelihood::Float64

chain::Vector{Float64}

expected_information::Float64

assessed::Bool

Factories

Latent1D(val, bounds, prior, posterior, chain, expected_information, assessed) = new(val, bounds, prior, posterior, chain, expected_information, assessed)

Creates a new 1-dimensional latent variable with custom fields.

Random Initializers

Latent1D()

Randomly generates a value for the 1-dimensional latent variable and assigns a default standardized Gaussian prior and posterior Latent1D(dist, bounds)

Same as Latent1D(dist, bounds) but with value equal to val Latent1D(val)

Randomly generates a value for the 1-dimensional latent variable and assigns a custom univariate distribution to prior and posterior with specific bounds.

Psychometrics._chain_append! — Method

_chain_append!(latent::Latent1D; sampling = false)

Psychometrics._empty_chain! — Method

_empty_chain!(latent::Latent1D)

Psychometrics._set_prior! — Method

_set_prior!(
    latent::Latent1D,
    prior::Union{Distributions.DiscreteUnivariateDistribution, Distributions.ContinuousUnivariateDistribution}
)

Psychometrics._set_val! — Method

_set_val!(latent::Latent1D, val::Float64)

Psychometrics._set_val_from_chain! — Method

_set_val_from_chain!(latent::Latent1D)

Psychometrics._update_estimate! — Method

_update_estimate!(latent::Latent1D; sampling = true)

Psychometrics.LatentND — Type

LatentND <: AbstractLatent

Description

A mutable describing a multivariate latent variable.

Fields

names::Vector{String}

val::Vector{Float64}

bounds::Vector{Vector{Float64}}

prior::Distributions.MultivariateDistribution

posterior::Distributions.MultivariateDistribution

chain::Vector{Vector{Float64}}

expected_information::Matrix{Float64}

Factories

LatentND(names, val, bounds, prior, posterior, chain, expectedinformation) = new(names, val, bounds, prior, posterior, chain, expectedinformation) Creates a new N-dimensional latent variable with custom fields.

Random Initializers

LatentND()

Randomly generates a value for the N-dimensional latent variable and assigns a default standardized Gaussian prior and posterior LatentND(dist, bounds)

Randomly generates a value for the N-dimensional latent variable and assigns a custom multivariate distribution to prior and posterior with specific bounds.