Examinees

Psychometrics.Examinee — Type

Examinee <: AbstractExaminee

Description

Examinee struct with a generic latent variable.

Fields

idx::Int64: An integer that identifies the Examinee in this session.

id::String: A string that identifies the Examinee.

latent::Latent: A generic latent variable associated with the Examinee.

Factories

Examinee1D(idx, id, latent) = new(idx, id, latent)

Creates a new Examinee with custom index, id and a generic latent variable.

Random initializers

Examinee(idx, id) = Examinee1D(idx, id)

Randomly generates an Examinee with custom index and id and with a default 1-dimensional latent variable (Look at (Latent1D)[#Psychometrics.Latent1D] for the defaults).

Psychometrics.Examinee1D — Type

Examinee1D <: AbstractExaminee

Description

Examinee struct with a 1-dimensional latent variable.

Fields

idx::Int64: An integer that identifies the Examinee in this session.

id::String: A string that identifies the Examinee.

latent::Latent1D: A 1-dimensional latent variable associated with the Examinee.

Factories

Examinee1D(idx, id, latent) = new(idx, id, latent)

Creates a new Examinee with custom index, id and 1-dimensional latent variable.

Random initializers

Examinee1D(idx, id) = new(idx, id, Latent1D())

Randomly generates an Examinee with custom index and id and with a default 1-dimensional latent variable (Look at (Latent1D)[#Psychometrics.Latent1D] for the defaults).

Psychometrics.get_examinee_by_id — Method

get_examinee_by_id(examinee_id::String, examinees::Vector{<:AbstractExaminee})

It returns the Examinee with index examinee_id from a Vector of AbstractExaminee.

Psychometrics.get_latents — Method

get_latents(examinees::Vector{<:AbstractExaminee})

Returns a matrix with latent values displayed by row.

Psychometrics.get_latents — Method

get_latents(examinee::Examinee1D)