Fisher Information

Psychometrics._item_expected_information — Method

_item_expected_information(parameters::Parameters1PL, latent::Latent1D)

Description

It computes the expected information (-second derivative of the likelihood) with respect to the difficulty parameter of the 1PL model. It follows the parametrization $a(θ - b)$.

Arguments

parameters::Parameters1PL : Required. A 1-parameter logistic parameters object.

latent::Latent1D : Required. A 1-dimensional Latent1D latent variable.

Output

A Float64 scalar.

Psychometrics._item_expected_information — Method

_item_expected_information(parameters::Parameters2PL, latent::Latent1D)

Description

It computes the expected information (-second derivative of the likelihood) with respect to the 2 parameters of the 2PL model. It follows the parametrization $a(θ - b)$.

Arguments

parameters::Parameters1PL : Required. A 2-parameter logistic parameters object.

latent::Latent1D : Required. A 1-dimensional Latent1D latent variable.

Output

A $2 \times 2$ matrix of the expected informations.

Psychometrics._item_expected_information — Method

_item_expected_information(parameters::Parameters3PL, latent::Latent1D)

Description

It computes the expected information (-second derivative of the likelihood) with respect to the 3 parameters of the 3PL model. It follows the parametrization $a(θ - b)$. It should be always positive.

Arguments

parameters::Parameters1PL : Required. A 3-parameter logistic parameters object.

latent::Latent1D : Required. A 1-dimensional Latent1D latent variable.

Output

A $3 \times 3$ matrix of the expected informations.

Psychometrics._item_observed_information — Method

_item_observed_information(
    parameters::Parameters1PL,
    latent::Latent1D,
    response_val::Float64,
)

Description

It is equal to the item expected information.

Arguments

parameters::Parameters1PL : Required. A 1-parameter logistic parameters object.

latent::Latent1D : Required. A 1-dimensional Latent1D latent variable.

response_val::Float64 : Required. A scalar response.

Output

A Float64 scalar observed (expected) informations.

Psychometrics._item_observed_information — Method

_item_observed_information(
    parameters::Parameters2PL,
    latent::Latent1D,
    response_val::Float64,
)

Description

It is equal to the item expected information.

Arguments

parameters::Parameters2PL : Required. A 2-parameter logistic parameters object.

latent::Latent1D : Required. A 1-dimensional Latent1D latent variable.

response_val::Float64 : Required. A scalar response.

Output

A $2 \times 2$ matrix of the observed (expected) informations.

Psychometrics._item_observed_information — Method

_item_observed_information(
    parameters::Parameters3PL,
    latent::Latent1D,
    response_val::Float64,
)

Description

It computes the observed information (-second derivative of the likelihood) with respect to the 3 parameters of the 3PL model. It follows the parametrization $a(θ - b)$.

Arguments

parameters::Parameters3PL : Required. A 3-parameter logistic parameters object.

latent::Latent1D : Required. A 1-dimensional Latent1D latent variable.

response_val::Float64 : Required. A scalar response.

Output

A $3 \times 3$ matrix of the observed informations.

Psychometrics._latent_information — Method

_latent_information(latent::Latent1D, parameters::Parameters1PL)

Description

It computes the information (second derivative of the likelihood) with respect to the 1-dimensional latent variable under the 1PL model. It follows the parametrization $a(θ - b)$.

Arguments

latent::Latent1D : Required. A 1-dimensional Latent1D latent variable.

parameters::Parameters1PL : Required. A 1-parameter logistic parameters object.

Output

A Float64 scalar.

Psychometrics._latent_information — Method

_latent_information(latent::Latent1D, parameters::Parameters2PL)

Description

It computes the information (-second derivative of the likelihood) with respect to the 1-dimensional latent variable under the 2PL model. It follows the parametrization $a(θ - b)$.

Arguments

latent::Latent1D : Required. A 1-dimensional Latent1D latent variable.

parameters::Parameters2PL : Required. A 2-parameter logistic parameters object.

Output

A Float64 scalar.

Psychometrics._latent_information — Method

_latent_information(latent::Latent1D, parameters::Parameters3PL)

Description

It computes the information (-second derivative of the likelihood) with respect to the 1-dimensional latent variable under the 3PL model. It follows the parametrization $a(θ - b)$.

Arguments

latent::Latent1D : Required. A 1-dimensional Latent1D latent variable.

parameters::Parameters3PL : Required. A 3-parameter logistic parameters object.

Output

A Float64 scalar.

Psychometrics.item_expected_information — Method

item_expected_information(
    item::AbstractItem,
    examinee::AbstractExaminee,
)

Description

Abstraction of itemexpected_information(latent, parameters) on Vector{<:AbstractExaminee} and items::Vector{<:AbstractItem}. It follows the parametrization $a(θ - b)$.

Arguments

item::AbstractItem : Required.

examinee::AbstractExaminee : Required.

Output

A matrix of expected informations (matrices for 2PL and 3PL or scalars 1PL).

Psychometrics.item_expected_information — Method

item_expected_information(
    items::Vector{<:AbstractItem},
    examinees::Vector{<:AbstractExaminee},
)

Description

Abstraction of itemexpected_information(latent, parameters) on Vector{<:AbstractExaminee} and items::Vector{<:AbstractItem}. It follows the parametrization $a(θ - b)$.

Arguments

items::Vector{<:AbstractItem} : Required. Size I x 1.

examinees::Vector{<:AbstractExaminee} : Required. Size N x 1.

Output

A $N \times I$ matrix of expected informations (matrices or scalars).

```@example examinee = Examinee(1); # default examinee random factory (1-D latent) item1PL = Item(1, Parameters1PL()); # default 1PL item random factory response1PL = answer(examinee, item1PL) # generate responses itemexpinfo1PL = itemexpectedinformation(item_1PL, examinee)

item2PL = Item(1, Parameters2PL()); # default 2PL item random factory response2PL = answer(examinee, item2PL) # generate responses itemexpinfo2PL = itemexpectedinformation(item_2PL, examinee)

Psychometrics.item_observed_information — Method

item_observed_information(
    item::AbstractItem,
    examinee::AbstractExaminee,
    response::AbstractResponse,
)

Description

It computes the item observed information for an examinee who answered response .

examinee = Examinee(1); # default examinee random factory (1-D latent)
item_2PL = Item(1, Parameters2PL()); # default 2PL item random factory
response_2PL = answer(examinee, item_2PL) # generate responses
item_obs_info_2PL = item_observed_information(item_2PL, examinee, response_2PL)
item_exp_info_2PL = item_expected_information(item_2PL, examinee)
# for 2PL and 1PL item parameters, the observed information is equal to the expected information.
@assert item_obs_info_2PL == item_exp_info_2PL

examinee = Examinee(1); # default examinee random factory (1-D latent)
item_3PL = Item(1, Parameters3PL()); # default 3PL item random factory
response_3PL = answer(examinee, item_3PL) # generate responses
# for 3PL item parameters, the observed information is different from the expected information.
item_obs_info_3PL = item_observed_information(item_3PL, examinee, response_3PL)
item_exp_info_3PL = item_expected_information(item_3PL, examinee)

Psychometrics.item_observed_information — Method

item_observed_information(
    items::Vector{<:AbstractItem},
    examinees::Vector{<:AbstractExaminee},
    responses::Vector{<:AbstractResponse},
)

Description

It computes the sum of the item observed informations across the responses. Items (latents) must be of the same type.

examinees = [Examinee(n) for n = 1 : 100]; #default examinee factory (1-D latent)
items_2PL = [Item(i, Parameters2PL()) for i = 1 : 30]; #2PL item factory
items_3PL = [Item(i + 30, string(i + 30), Parameters3PL()) for i = 1 : 30]; #3PL item factory, id and idx must be different
items = vcat([items_2PL, items_3PL]...) # create vector of items

responses = answer(examinees, items) # generate responses
#obs_item_info = item_observed_information(items, examinees, responses) # do not run
#^ This returns an error since Julia cannot sum ``2 \times 2`` matrices with ``3 \times 3`` matrices.
# Instead do this ⌄.
obs_items_info_2PL = item_observed_information(items_2PL, examinees, responses)
obs_items_info_3PL = item_observed_information(items_3PL, examinees, responses)

Psychometrics.latent_information — Method

latent_information(examinee::AbstractExaminee, item::AbstractItem)

Description

An abstraction of _latent_information(latent::AbstractLatent, parameters::AbstractParametersBinary) on an examinee and an item. It follows the parametrization $a(θ - b)$.

Arguments

examinee::AbstractExaminee : Required.

item::AbstractItem : Required.

Output

A Float64 scalar.

Psychometrics.latent_information — Method

latent_information(examinee::AbstractExaminee, items::Vector{<:AbstractItem})

Description

An abstraction of _latent_information(latent::AbstractLatent, parameters::AbstractParametersBinary) on an examinee and items. It follows the parametrization $a(θ - b)$.

Arguments

examinee::AbstractExaminee : Required.

items::Vector{<:AbstractItem} : Required.

Output

A Float64 vector.

Example

Compute the Fisher information for the latent/s of the examinees (second derivatives of the likelihood with respect to $\theta$) and each item.

$E_\theta_n \[ I(\theta_n | b_i) \]$

examinee = Examinee(); #default examinee factory
items = [Item() for i = 1 : 30]; #default item factory
latent_information(examinee, items) #compute the information wrt θ   

Psychometrics.latent_information — Method

latent_information(
    latents_matrix::Matrix{Float64},
    parameters_matrix::Matrix{Float64},
)

Description

It computes the information function (IIF) for item parameters at latents values provided in matrix form. Not suitable for 3PL models, for such a kind of model use latent_information_3PL(). It follows the parametrization $a \theta - b$. See the docs of latent_information(examinee::AbstractExaminee, item::AbstractItem) for details.

Arguments

latents_matrix::Matrix{Float64} : Required. A $n_latents \times N$ matrix with latents values.

parameters_matrix::Matrix{Float64} : Required. A $(n_latents + 1) \times I$ matrix with item parameters. intercept (b) must be in first row, latents coefficients $(a_j)$ in next rows $(2, \ldots, n_latents + 1)$.

Output

A $I \time N$ Float64 matrix.

Psychometrics.latent_information — Method

latent_information(
    examinees::Vector{<:AbstractExaminee},
    items::Vector{<:AbstractItem},
)

Description

An abstraction of _latent_information(latent::AbstractLatent, parameters::AbstractParametersBinary) on examinees and items. It follows the parametrization $a(θ - b)$.

Arguments

examinees::Vector{<:AbstractExaminee} : Required.

items::Vector{<:AbstractItem} : Required.

Output

A $N \times I$ matrix of latent informations (matrices or scalars).

examinees = [Examinee(); for n = 1 : 100]; #default examinee factory
items = [Item() for i = 1 : 30]; #default item factory
latent_information(examinees, items) #compute the information wrt θs   

Psychometrics.latent_information_3PL — Method

latent_information_3PL(
    latents_matrix::Matrix{Float64},
    parameters_matrix::Matrix{Float64},
)

Description

Only for models which has guessing parameter (c) in last row of parameters_matrix. It computes the information function (IIF) for item parameters at latents values provided in matrix form. It follows the parametrization $a \theta - b$.

Arguments

latents_matrix::Matrix{Float64} : Required. A n_latents x N matrix with latents values.

parameters_matrix::Matrix{Float64} : Required. A (n_latents + 1) x I matrix with item parameters. intercept (b) must be in first row, latents coefficients (aj) in next rows (2, ..., nlatents + 1).

Output

A I x N Float64 matrix.