Items

Psychometrics.Item — Type

Item <: AbstractItem

Description

A generic item struct.

Fields

idx::Int64: An integer that identifies the item in this session.

id::String: A string that identifies the examinee.

content::Vector{String}: A string vector containing the content features of an item.

parameters::AbstractParameters: A generic item parameters object.

calibrated::Bool: Tells if the item has been already calibrated.

Factories

Item(idx, id, content, parameters) = new(idx, id, content, parameters)

Creates a new generic item with custom index, id, content features and item parameters.

Random initilizers

Item(idx, id, content) = new(idx, id, content, Parameters1PL(), true)

Randomly generates a new generic calibrated item with custom index, id, content features and default 1PL item parameters (Look at (Parameters1PL)[#Psychometrics.Parameters1PL] for the defaults).

Psychometrics.chain_append! — Method

chain_append!(item::AbstractItem; sampling = false)

Psychometrics.empty_chain! — Method

empty_chain!(item::AbstractItem)

Psychometrics.get_item_by_id — Method

get_item_by_id(item_id::String, items::Vector{<:AbstractItem})

It returns the item with index item_id from a vector of <:AbstractItem.

Psychometrics.get_parameters — Method

get_parameters(item::AbstractItem)

Psychometrics.get_parameters — Method

get_parameters(items::Vector{<:AbstractItem})

Returns a matrix with item parameters displayed by row.

Psychometrics._add_posterior! — Method

_add_posterior!(parameters::AbstractParameters, priors::Vector{Distributions.UnivariateDistribution})

Description

It transforms the vector posteriors of univariate distributions to their products and assign it to AbstractParameters instance.

Arguments

parameters::AbstractParameters : Required. Any type of parameters object.

posteriors::Vector{Distributions.UnivariateDistribution} : Required. A vector of probability distributions. The size of the vector must be the same as the number of parameters.

Examples

parameters2PL = Parameters2PL()
a_dist = Distributions.Normal(0,1)
b_dist = Distributions.Normal(0,1)
_add_posterior!(parameters2PL, [a_dist, b_dist])

Psychometrics._add_posterior! — Method

_add_posterior!(parameters::AbstractParameters, posterior::Distributions.Distribution)

Description

It assigns the <n>-variate posterior distribution to a AbstractParameters instance with <n> parameters.

Arguments

parameters::AbstractParameters : Required. Any type of parameters object.

posterior::Distributions.Distribution : Required. A <n>-variate probability distribution where <n> > 1 and is the numebr of item parameters in parameters.

Examples

parameters2PL = Parameters2PL()
bivariate_normal = Distributions.MultivariateNormal([0,0], LinearAlgebra.I(2))
_add_posterior!(parameters2PL, bivariate_normal)

Psychometrics.add_prior! — Method

_add_prior!(parameters::AbstractParameters, priors::Vector{Distributions.UnivariateDistribution})

Description

It transforms the vector priors of univariate distributions to their products and assign it to AbstractParameters instance.

Arguments

parameters::AbstractParameters : Required. Any type of parameters object.

priors::Vector{Distributions.UnivariateDistribution} : Required. A vector of probability distributions. The size of the vector must be the same as the number of item parameters.

Examples

parameters2PL = Parameters2PL()
a_dist = Distributions.Normal(0,1)
b_dist = Distributions.Normal(0,1)
_add_prior!(parameters2PL, [a_dist, b_dist])

Psychometrics.add_prior! — Method

_add_prior!(parameters::AbstractParameters, prior::Distributions.Distribution)

Description

It assigns the prior prior to a AbstractParameters instance.

Arguments

parameters::AbstractParameters : Required. Any type of parameters object.

prior::Distributions.Distribution : Required. A <n>-variate probability distribution where <n> > 1 and is the numebr of item parameters in parameters.

Examples

parameters2PL = Parameters2PL()
bivariate_normal = Distributions.MultivariateNormal([0,0], LinearAlgebra.I(2))
_add_prior!(parameters2PL, bivariate_normal)

Psychometrics.Parameters1PL — Type

Parameters1PL <: AbstractParametersBinary

Contains information of a set of item parameters (values, bounds, priors, posteiors, chains, expected Fisher information, calibrated) under the 1-parameter logistic model.

Psychometrics._empty_chain! — Method

_empty_chain!(parameters::Parameters1PL)

Psychometrics.Parameters2PL — Type

Parameters2PL <: AbstractParametersBinary

Contains information of a set of item parameters (values, bounds, priors, posteiors, chains, expected Fisher information, calibrated) under the 2-parameter logistic model.

Psychometrics._empty_chain! — Method

_empty_chain!(parameters::Parameters2PL)

Psychometrics._set_val! — Method

_set_val!(parameters::Parameters2PL, vals::Vector{Float64})

Psychometrics._set_val_from_chain! — Method

_set_val_from_chain!(parameters::Parameters2PL)

Psychometrics._update_estimate! — Method

_update_estimate!(parameters::Parameters2PL; sampling = true)

Psychometrics.Parameters3PL — Type

Parameters3PL <: AbstractParametersBinary

Contains information of a set of item parameters (values, bounds, priors, posteiors, chains, expected Fisher information, calibrated) under the 3-parameter logistic model.

Psychometrics._chain_append! — Method

_chain_append!(parameters::Union{Parameters2PL,Parameters3PL}; sampling = false)

Psychometrics._empty_chain! — Method

_empty_chain!(parameters::Parameters3PL)