Items

Psychometrics.AbstractItem — Type

AbstractItem

Psychometrics.Item — Type

Item <: AbstractItem

Description

An immutable containing information about an item (a question in a test), e.g. id::String the item identifier, and the field parameters::AbstractParameter which accepts a mutable item parameter object

Fields

idx::Int64: An integer that identifies the item in this session.

id::String: A string that identifies the examinee.

content::Vector{String}: A string vector containing the content features of an item.

parameters::AbstractParameters: A generic item parameters object.

Factories

Item(idx::Int64, id::String, content::Vector{String}, parameters::AbstractParameters) = new(idx, id, content, parameters)
Item(idx::Int64, id::String, parameters::AbstractParameters) = new(idx, id, String[], parameters)
Item(idx::Int64, parameters::AbstractParameters) = new(idx, string(idx), String[], parameters)

Creates a new generic item with custom index, id, (content features) and item parameters.

Random initilizers

Item(idx::Int64, id::String) = new(idx, id, String[], Parameters1PL())
Item(idx::Int64, id::String, content::String) = new(idx, id, content, Parameters1PL())

Randomly generates a new generic calibrated item with custom index, id, content features and default 1PL item parameters (Look at (Parameters1PL)[#Psychometrics.Parameters1PL] for the defaults).

Psychometrics.chain_append! — Method

chain_append!(item::AbstractItem; sampling = false)

Psychometrics.empty_chain! — Method

empty_chain!(item::AbstractItem)

Psychometrics.get_item_by_id — Method

get_item_by_id(item_id::String, items::Vector{<:AbstractItem})

Description

It returns the item with index item_id from a vector of <:AbstractItem.

Psychometrics.get_parameters — Method

get_parameters(item::AbstractItem)

Psychometrics.get_parameters_vals — Method

get_parameters_vals(item::AbstractItem)

Psychometrics.AbstractParameters — Type

Description

An abstract type which, at the moment, has only the abstract AbstractParametersBinary as subtype. Thus, it works only with binary (dichotomous) responses. The available mutables for the latter are Parameters1PL, Parameters2PL, Parameters3PL. They contain the details about the item parameters under the 1-parameter logistic (1PL), 2-parameters logistic (2PL) and 3-parameter logistic (3PL) IRT models, respectively.

Example

The object Parameters2PL has the fields difficulty b::Float64 and discrimination a::Float64. It is possible to define the Bayesian priors and posterior by assigning multivariate distributions from the package Distributions to the fields prior::Distributions.MultivariateDistribution and posterior::Distributions.MultivariateDistribution.

Psychometrics._add_posterior! — Method

_add_posterior!(
    parameters::AbstractParameters,
    posterior::Distributions.Distribution,
)

Description

It assigns the <n>-variate posterior distribution to a AbstractParameters instance with <n> parameters.

Arguments

parameters::AbstractParameters : Required. Any type of parameters object.

posterior::Distributions.Distribution : Required. A <n>-variate probability distribution where <n> > 1 and is the numebr of item parameters in parameters.

Examples

parameters2PL = Parameters2PL()
bivariate_normal = Distributions.MultivariateNormal([0,0], LinearAlgebra.I(2))
_add_posterior!(parameters2PL, bivariate_normal)

Psychometrics._add_posterior! — Method

_add_posterior!(
    parameters::AbstractParameters,
    posteriors::Vector{Distributions.UnivariateDistribution},
)

Description

It transforms the vector posteriors of univariate distributions to their products and assign it to AbstractParameters instance.

Arguments

parameters::AbstractParameters : Required. Any type of parameters object.

posteriors::Vector{Distributions.UnivariateDistribution} : Required. A vector of probability distributions. The size of the vector must be the same as the number of parameters.

Examples

parameters2PL = Parameters2PL()
a_dist = Distributions.Normal(0,1)
b_dist = Distributions.Normal(0,1)
_add_posterior!(parameters2PL, [a_dist, b_dist])

Psychometrics.add_prior! — Method

add_prior!(
    parameters::AbstractParameters,
    prior::Distributions.MultivariateDistribution,
)

Description

It assigns the prior prior to a AbstractParameters instance.

Arguments

parameters::AbstractParameters : Required. Any type of parameters object.

prior::Distributions.Distribution : Required. A <n>-variate probability distribution where <n> > 1 and is the numebr of item parameters in parameters.

Examples

parameters2PL = Parameters2PL()
bivariate_normal = Distributions.MultivariateNormal([0,0], LinearAlgebra.I(2))
_add_prior!(parameters2PL, bivariate_normal)

Psychometrics.add_prior! — Method

add_prior!(
    parameters::AbstractParameters,
    priors::Vector{Distributions.UnivariateDistribution},
)

Description

It transforms the vector priors of univariate distributions to their products and assign it to AbstractParameters instance.

Arguments

parameters::AbstractParameters : Required. Any type of parameters object.

priors::Vector{Distributions.UnivariateDistribution} : Required. A vector of probability distributions. The size of the vector must be the same as the number of item parameters.

Examples

parameters2PL = Parameters2PL()
a_dist = Distributions.Normal(0,1)
b_dist = Distributions.Normal(0,1)
_add_prior!(parameters2PL, [a_dist, b_dist])

Psychometrics.Parameters1PL — Type

Parameters1PL <: AbstractParametersBinary

Contains information of a set of item parameters (values, bounds, priors, posteiors, chains, expected Fisher information, calibrated) under the 1-parameter logistic model.

Psychometrics._empty_chain! — Method

_empty_chain!(parameters::Parameters1PL)

Psychometrics.Parameters2PL — Type

Parameters2PL <: AbstractParametersBinary

Contains information of a set of item parameters (values, bounds, priors, posteiors, chains, expected Fisher information, calibrated) under the 2-parameter logistic model.

Psychometrics._empty_chain! — Method

_empty_chain!(parameters::Parameters2PL)

Psychometrics._get_parameters_vals — Method

_get_parameters_vals(parameters::Parameters2PL)

Psychometrics._set_val! — Method

_set_val!(parameters::Parameters2PL, vals::Vector{Float64})

Psychometrics._set_val_from_chain! — Method

_set_val_from_chain!(parameters::Parameters2PL)

Psychometrics._update_estimate! — Method

_update_estimate!(parameters::Parameters2PL; sampling = true)

Psychometrics.Parameters3PL — Type

Parameters3PL <: AbstractParametersBinary

Contains information of a set of item parameters (values, bounds, priors, posteiors, chains, expected Fisher information, calibrated) under the 3-parameter logistic model.

Psychometrics._chain_append! — Method

_chain_append!(parameters::Union{Parameters2PL,Parameters3PL}; sampling = false)

Psychometrics._chain_append_and_set_val! — Method

_chain_append_and_set_val!(parameters::Parameters2PL; sampling = false)

Psychometrics._chain_append_and_set_val! — Method

_chain_append_and_set_val!(parameters::Parameters3PL; sampling = false)

Psychometrics._empty_chain! — Method

_empty_chain!(parameters::Parameters3PL)