Log Likelihood

Psychometrics.__likelihood — Method

__likelihood(
    response_val::Float64,
    latent_val::Float64,
    parameters::AbstractParametersBinary;
    weight::Float64 = 1.0
)

#Description

It computes the likelihood for a latent value and item parameters parameters with answer value response_val. Optionally weights the result by setting weight, that is equal to 1.0 by default.

Psychometrics._likelihood — Method

_likelihood(
    response::AbstractResponse,
    latent_val::Float64,
    item::AbstractItem;
    weight::Float64 = 1.0
)

#Description

It computes the likelihood for a latent value and item item with answer response. Optionally weights the result by setting weight, that is equal to 1.0 by default.

Psychometrics._likelihood — Method

_likelihood(
    response_val::Float64,
    latent::Latent1D,
    parameters::AbstractParametersBinary,
)

Description

It computes the likelihood for a 1-dimensional latent variable and item parameters parameters with answer response_val.

Psychometrics._log_likelihood — Method

_log_likelihood(
    response_val::Float64,
    latent::Latent1D,
    parameters::AbstractParametersBinary,
    g_item::Vector{Float64},
    g_latent::Vector{Float64},
)

Description

It computes the log likelihood for a 1-dimensional latent variable and item parameters parameters with answer response_val. It updates also the gradient vectors.

Psychometrics._log_likelihood — Method

_log_likelihood(
    response_val::Float64,
    latent::Latent1D,
    parameters::AbstractParametersBinary,
)

Description

It computes the log likelihood for a 1-dimensional latent variable and item parameters parameters with answer response_val.

Psychometrics.likelihood — Method

likelihood(
    response::AbstractResponse,
    examinee::AbstractExaminee,
    item::AbstractItem,
)

Description

It computes the likelihood for a triplet response, examinee, item.

Psychometrics.likelihood — Method

likelihood(
    responses::Vector{<:AbstractResponse},
    examinees::Vector{<:AbstractExaminee},
    items::Vector{<:AbstractItem},
)

Description

It computes the likelihood for a vector of responses responses, and provided examinees and items vectors.

Psychometrics.log_likelihood — Method

log_likelihood(
    response::AbstractResponse,
    examinee::AbstractExaminee,
    item::AbstractItem,
    g_item::Vector{Float64},
    g_latent::Vector{Float64},
)

Description

It computes the log likelihood for a triplet response, examinee, item. It updates also the gradient vectors.

Psychometrics.log_likelihood — Method

log_likelihood(
    response::AbstractResponse,
    examinee::AbstractExaminee,
    item::AbstractItem,
)

Description

It computes the log likelihood for a triplet response, examinee, item.

Psychometrics.log_likelihood — Method

log_likelihood(
    response_val::Float64,
    latent_val::Float64,
    parameters::AbstractParametersBinary,
)

Description

It computes the log likelihood for a latent value and item parameters parameters with answer response_val.

Psychometrics.log_likelihood — Method

log_likelihood(
    responses::Vector{<:AbstractResponse},
    examinees::Vector{<:AbstractExaminee},
    items::Vector{<:AbstractItem},
    g_item::Vector{Float64},
    g_latent::Vector{Float64},
)

Description

It computes the log likelihood for a vector of responses responses, and provided examinees and items vectors. It updates also the gradient vectors.

Psychometrics.log_likelihood — Method

log_likelihood(
    responses::Vector{<:AbstractResponse},
    examinees::Vector{<:AbstractExaminee},
    items::Vector{<:AbstractItem},
)

Description

It computes the log likelihood for a vector of responses responses, and provided examinees and items vectors.