Log Likelihood

Psychometrics.likelihood — Method

likelihood(response::AbstractResponse, examinee::AbstractExaminee, item::AbstractItem)

It computes the log likelihood for a response.

Psychometrics.likelihood — Method

likelihood(responses::Vector{<:AbstractResponse},  examinees::Vector{<:AbstractExaminee}, items::Vector{<:AbstractItem})

It computes the log likelihood for a vector of responses responses.

Psychometrics.likelihood — Method

likelihood(response_val::Float64, latent_val::Float64, parameters::AbstractParameters)

It computes the log likelihood for a latent value and item parameters parameters with answer response_val.

Psychometrics.likelihood — Method

likelihood(response_val::Float64, latent::Latent1D, parameters::AbstractParameters)

It computes the likelihood for a 1-dimensional latent variable and item parameters parameters with answer response_val.

Psychometrics.log_likelihood — Method

log_likelihood(response::AbstractResponse, examinee::AbstractExaminee, item::AbstractItem, g_item::Vector{Float64}, g_latent::Vector{Float64})

It computes the log likelihood for a response response. It updates also the gradient vectors.

Psychometrics.log_likelihood — Method

log_likelihood(response::AbstractResponse, examinee::AbstractExaminee, item::AbstractItem)

It computes the log likelihood for a response.

Psychometrics.log_likelihood — Method

log_likelihood(responses::Vector{<:AbstractResponse},  examinees::Vector{<:AbstractExaminee}, items::Vector{<:AbstractItem})

It computes the log likelihood for a vector of responses responses.

Psychometrics.log_likelihood — Method

log_likelihood(responses::Vector{<:AbstractResponse},  examinees::Vector{<:AbstractExaminee}, items::Vector{<:AbstractItem}, g_item::Vector{Float64}, g_latent::Vector{Float64})

It computes the log likelihood for a vector of responses responses. It updates also the gradient vectors.

Psychometrics.log_likelihood — Method

log_likelihood(responses::Matrix{Float64}, latents_matrix::Matrix{Float64}, parameters_matrix::Matrix{Float64}, design::Matrix{Float64})

It computes the log likelihood for a matrix of latent values, item parameters, responses and design.

Psychometrics.log_likelihood — Method

log_likelihood(response_val::Float64, latent::Latent1D, parameters::AbstractParameters, g_item::Vector{Float64}, g_latent::Vector{Float64})

It computes the log likelihood for a 1-dimensional latent variable and item parameters parameters with answer response_val. It updates also the gradient vectors.

Psychometrics.log_likelihood — Method

log_likelihood(response_val::Float64, latent::Latent1D, parameters::AbstractParameters)

It computes the log likelihood for a 1-dimensional latent variable and item parameters parameters with answer response_val.