Log Likelihood

Psychometrics.log_likelihood — Method

log_likelihood(response::AbstractResponse, examinee::AbstractExaminee, item::AbstractItem, g_item::Vector{Float64}, g_latent::Vector{Float64})

It computes the log likelihood for a response response. It updates also the gradient vectors.

Psychometrics.log_likelihood — Method

log_likelihood(response::AbstractResponse, examinee::AbstractExaminee, item::AbstractItem)

It computes the log likelihood for a response.

Psychometrics.log_likelihood — Method

log_likelihood(responses::Vector{<:AbstractResponse},  examinees::Dict{Int64,<:AbstractExaminee}, items::Dict{Int64,<:AbstractItem}, g_item::Vector{Float64}, g_latent::Vector{Float64})

It computes the log likelihood for a vector of responses responses. It updates also the gradient vectors.

Psychometrics.log_likelihood — Method

log_likelihood(responses::Vector{<:AbstractResponse},  examinees::Dict{Int64,<:AbstractExaminee}, items::Dict{Int64,<:AbstractItem})

It computes the log likelihood for a vector of responses responses.

Psychometrics.log_likelihood — Method

log_likelihood(response_val::Float64, latent::Latent1D, parameters::AbstractParameters, g_item::Vector{Float64}, g_latent::Vector{Float64})

It computes the log likelihood for a 1-dimensional latent variable and item parameters parameters with answer response_val. It updates also the gradient vectors.

Psychometrics.log_likelihood — Method

log_likelihood(response_val::Float64, latent::Latent1D, parameters::AbstractParameters)

It computes the log likelihood for a 1-dimensional latent variable and item parameters parameters with answer response_val.