Information (Item Information Function)

Information wrt $\theta$

Values in Matrix form

Missing docstring.

Missing docstring for latent_information(latents_matrix::Matrix{Float64}, parameters_matrix::Matrix{Float64}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for latent_information_3PL( latents_matrix::Matrix{Float64}, parameters_matrix::Matrix{Float64}, ). Check Documenter's build log for details.

Using Structs

Missing docstring.

Missing docstring for latent_information(examinee::AbstractExaminee, item::AbstractItem). Check Documenter's build log for details.

Psychometrics.latent_information — Method

latent_information(examinee::AbstractExaminee, items::Vector{<:AbstractItem})

Description

An abstraction of _latent_information(latent::AbstractLatent, parameters::AbstractParametersBinary) on an examinee and items. It follows the parametrization $a(θ - b)$.

Arguments

examinee::AbstractExaminee : Required.

items::Vector{<:AbstractItem} : Required.

Output

A Float64 vector.

Example

Compute the Fisher information for the latent/s of the examinees (second derivatives of the likelihood with respect to $\theta$) and each item.

$E_\theta_n \[ I(\theta_n | b_i) \]$

examinee = Examinee(); #default examinee factory
items = [Item() for i = 1 : 30]; #default item factory
latent_information(examinee, items) #compute the information wrt θ   

Psychometrics.latent_information — Method

latent_information(
    examinees::Vector{<:AbstractExaminee},
    items::Vector{<:AbstractItem},
)

Description

An abstraction of _latent_information(latent::AbstractLatent, parameters::AbstractParametersBinary) on examinees and items. It follows the parametrization $a(θ - b)$.

Arguments

examinees::Vector{<:AbstractExaminee} : Required.

items::Vector{<:AbstractItem} : Required.

Output

A $N \times I$ matrix of latent informations (matrices or scalars).

examinees = [Examinee(); for n = 1 : 100]; #default examinee factory
items = [Item() for i = 1 : 30]; #default item factory
latent_information(examinees, items) #compute the information wrt θs   

Information wrt Item Parameters

Psychometrics.item_expected_information — Method

item_expected_information(
    items::Vector{<:AbstractItem},
    examinees::Vector{<:AbstractExaminee},
)

Description

Abstraction of itemexpected_information(latent, parameters) on Vector{<:AbstractExaminee} and items::Vector{<:AbstractItem}. It follows the parametrization $a(θ - b)$.

Arguments

items::Vector{<:AbstractItem} : Required. Size I x 1.

examinees::Vector{<:AbstractExaminee} : Required. Size N x 1.

Output

A $N \times I$ matrix of expected informations (matrices or scalars).

```@example examinee = Examinee(1); # default examinee random factory (1-D latent) item1PL = Item(1, Parameters1PL()); # default 1PL item random factory response1PL = answer(examinee, item1PL) # generate responses itemexpinfo1PL = itemexpectedinformation(item_1PL, examinee)

item2PL = Item(1, Parameters2PL()); # default 2PL item random factory response2PL = answer(examinee, item2PL) # generate responses itemexpinfo2PL = itemexpectedinformation(item_2PL, examinee)

Psychometrics.item_observed_information — Method

item_observed_information(
    items::Vector{<:AbstractItem},
    examinees::Vector{<:AbstractExaminee},
    responses::Vector{<:AbstractResponse},
)

Description

It computes the sum of the item observed informations across the responses. Items (latents) must be of the same type.

examinees = [Examinee(n) for n = 1 : 100]; #default examinee factory (1-D latent)
items_2PL = [Item(i, Parameters2PL()) for i = 1 : 30]; #2PL item factory
items_3PL = [Item(i + 30, string(i + 30), Parameters3PL()) for i = 1 : 30]; #3PL item factory, id and idx must be different
items = vcat([items_2PL, items_3PL]...) # create vector of items

responses = answer(examinees, items) # generate responses
#obs_item_info = item_observed_information(items, examinees, responses) # do not run
#^ This returns an error since Julia cannot sum ``2 \times 2`` matrices with ``3 \times 3`` matrices.
# Instead do this ⌄.
obs_items_info_2PL = item_observed_information(items_2PL, examinees, responses)
obs_items_info_3PL = item_observed_information(items_3PL, examinees, responses)