Likelihood

Values in Matrix form

Psychometrics.log_likelihood — Method

loglikelihood( responseval::Float64, latent_val::Float64, parameters::AbstractParametersBinary, )

Description

It computes the log likelihood for a matrix of latent values, item parameters, responses and design.

Using Structs

Missing docstring.

Missing docstring for log_likelihood(responses::Matrix{Float64}, latents_matrix::Matrix{Float64}, parameters_matrix::Matrix{Float64}, design::Matrix{Float64}). Check Documenter's build log for details.

Psychometrics.log_likelihood — Method

log_likelihood(response::AbstractResponse, examinee::AbstractExaminee, item::AbstractItem)

It computes the log likelihood for a response.

Psychometrics.log_likelihood — Method

log_likelihood(response::AbstractResponse, examinee::AbstractExaminee, item::AbstractItem, g_item::Vector{Float64}, g_latent::Vector{Float64})

It computes the log likelihood for a response response. It updates also the gradient vectors.

Psychometrics.log_likelihood — Method

log_likelihood(responses::Vector{<:AbstractResponse},  examinees::Vector{<:AbstractExaminee}, items::Vector{<:AbstractItem})

It computes the log likelihood for a vector of responses responses.

Psychometrics.log_likelihood — Method

log_likelihood(responses::Vector{<:AbstractResponse},  examinees::Vector{<:AbstractExaminee}, items::Vector{<:AbstractItem}, g_item::Vector{Float64}, g_latent::Vector{Float64})

It computes the log likelihood for a vector of responses responses. It updates also the gradient vectors.

Psychometrics.likelihood — Method

likelihood(response::AbstractResponse, examinee::AbstractExaminee, item::AbstractItem)

It computes the log likelihood for a response.

Psychometrics.likelihood — Method

likelihood(responses::Vector{<:AbstractResponse},  examinees::Vector{<:AbstractExaminee}, items::Vector{<:AbstractItem})

It computes the log likelihood for a vector of responses responses.