Likelihood

Missing docstring.

Missing docstring for log_likelihood( response_val::Float64, latent_val::Float64, parameters::AbstractParametersBinary, ). Check Documenter's build log for details.

Missing docstring.

Missing docstring for log_likelihood( response::AbstractResponse, examinee::AbstractExaminee, item::AbstractItem, ). Check Documenter's build log for details.

Missing docstring.

Missing docstring for log_likelihood( response::AbstractResponse, examinee::AbstractExaminee, item::AbstractItem, g_item::Vector{Float64}, g_latent::Vector{Float64}, ). Check Documenter's build log for details.

Psychometrics.log_likelihood — Method

log_likelihood(
    responses::Vector{<:AbstractResponse},
    examinees::Vector{<:AbstractExaminee},
    items::Vector{<:AbstractItem},
)

Description

It computes the log likelihood for a vector of responses responses, and provided examinees and items vectors.

Psychometrics.log_likelihood — Method

log_likelihood(
    responses::Vector{<:AbstractResponse},
    examinees::Vector{<:AbstractExaminee},
    items::Vector{<:AbstractItem},
    g_item::Vector{Float64},
    g_latent::Vector{Float64},
)

Description

It computes the log likelihood for a vector of responses responses, and provided examinees and items vectors. It updates also the gradient vectors.

Missing docstring.

Missing docstring for likelihood( response::AbstractResponse, examinee::AbstractExaminee, item::AbstractItem, ). Check Documenter's build log for details.

Psychometrics.likelihood — Method

likelihood(
    responses::Vector{<:AbstractResponse},
    examinees::Vector{<:AbstractExaminee},
    items::Vector{<:AbstractItem},
)

Description

It computes the likelihood for a vector of responses responses, and provided examinees and items vectors.