Probability (Item Characteristic Function)

Values in Matrix form

Psychometrics.probability — Method

probability(parameters_matrix::Matrix{Float64}, latent_matrix::Matrix{Float64})

Description

It computes the probability (ICF) of a correct response for item parameters and latents values provided in matrix form. Not suitable for 3PL models, for such a kind of model use probability_3PL(). It follows the parametrization $a \theta - b$.

Arguments

parameters_matrix::Matrix{Float64} : Required. A I x (n_latents + 1) matrix with item parameters. intercept (b) must be in first column, latents coefficients (aj) in next columns (2, ..., nlatents + 1).

latent_matrix::Matrix{Float64} : Required. A n_latents x N or (n_latents + 1) x N matrix with latents values.

Output

A I x N Float64 matrix.

Psychometrics.probability_3PL — Method

probability_3PL(parameters_matrix::Matrix{Float64}, latent_matrix::Matrix{Float64})

Description

Only for models which has guessing parameter (c) in last row of parameters_matrix. It computes the probability (ICF) of a correct response for item parameters and latents values provided in matrix form. It follows the parametrization $a \theta - b$.

Arguments

parameters_matrix::Matrix{Float64} : Required. A I x (n_latents + 1) matrix with item parameters. intercept (b) must be in first column, latents coefficients (aj) in next columns (2, ..., nlatents + 1).

latent_matrix::Matrix{Float64} : Required. A n_latents x N matrix with latents values.

Output

A I x N Float64 matrix.

Using Structs

Psychometrics.probability — Method

probability(examinee::AbstractExaminee, item::AbstractItem)

Description

It computes the probability (ICF) that an examinee answers correctly at item.

Arguments

examinee::AbstractExaminee : Required. An Examinee.

item::AbstractItem : Required. An Item.

Output

A Float64 scalar.

Psychometrics.probability — Method

probability(examinees::Vector{<:AbstractExaminee}, items::Vector{<:AbstractItem})

Description

It computes the probability (ICF) that a vector of examinees answers correctly at a vector of items.

Arguments

examinees::Vector{<:AbstractExaminee} : Required. A vector of Examinees.

items::Vector{<:AbstractItem} : Required. A vector of Items.

Output

A I x N Float64 matrix.

Psychometrics.probability — Method

probability(items::Vector{<:AbstractItem}, examinees::Vector{<:AbstractExaminee})

Description

It computes the probability (ICF) that a vector of examinees answers correctly at a vector of items.

Arguments

examinees::Vector{<:AbstractExaminee} : Required. A vector of Examinees.

items::Vector{<:AbstractItem} : Required. A vector of Items.

Output

A I x N Float64 matrix.