Likelihood (Internals)

Using Structs

Psychometrics._likelihood — Method

_likelihood(
    response_val::Float64,
    latent_val::Float64,
    parameters::AbstractParametersBinary,
)

#Description

It computes the likelihood for a latent value and item parameters parameters with answer response_val.

Psychometrics._log_likelihood — Method

_log_likelihood(
    response_val::Float64,
    latent::Latent1D,
    parameters::AbstractParametersBinary,
)

Description

It computes the log likelihood for a 1-dimensional latent variable and item parameters parameters with answer response_val.

Psychometrics._log_likelihood — Method

_log_likelihood(
    response_val::Float64,
    latent::Latent1D,
    parameters::AbstractParametersBinary,
    g_item::Vector{Float64},
    g_latent::Vector{Float64},
)

Description

It computes the log likelihood for a 1-dimensional latent variable and item parameters parameters with answer response_val. It updates also the gradient vectors.