Probability (Item Characteristic Function)

Psychometrics.__probability — Method

__probability(latent_vals::Vector{Float64}, parameters::Parameters3PL)

Description

It computes the probability (ICF) of a correct response for item parameters under the 2PL model at latent_vals points. N-dimensional latent. It follows the parametrization $a(θ - b)$.

Arguments

latent_vals::Vector{Float64} : Required. The points in the latent space in which compute the probability.

parameters::ParametersNPL : Required. A 2-parameter logistic parameters object suitbale for N-dimensional latent.

Output

A Float64 scalar.

Psychometrics.__probability — Method

__probability(latent_val::Float64, parameters::Parameters1PL)

Description

It computes the probability (ICF) of a correct response for item parameters under the 1PL model at latent_val point. It follows the parametrization $a(θ - b)$.

Arguments

latent_val::Float64 : Required. The point in the latent space in which compute the probability.

parameters::Parameters1PL : Required. A 1-parameter logistic parameters object.

Output

A Float64 scalar.

Psychometrics.__probability — Method

 __probability(latent_val::Float64, parameters::Parameters2PL)

Description

It computes the probability (ICF) of a correct response for item parameters under the 2PL model at latent_val point. It follows the parametrization $a(θ - b)$.

Arguments

latent_val::Float64 : Required. The point in the latent space in which compute the probability.

parameters::Parameters2PL : Required. A 2-parameter logistic parameters object.

Output

A Float64 scalar.

Psychometrics.__probability — Method

__probability(latent_val::Float64, parameters::Parameters3PL)

Description

It computes the probability (ICF) of a correct response for item parameters under the 3PL model at latent_val point. It follows the parametrization $a(θ - b)$.

Arguments

latent_val::Float64 : Required. The point in the latent space in which compute the probability.

parameters::Parameters3PL : Required. A 3-parameter logistic parameters object.

Output

A Float64 scalar.

Psychometrics._probability — Method

_probability(latent::Latent1D, parameters::Parameters1PL, g_item::Vector{Float64}, g_latent::Vector{Float64})

Description

It computes the probability (ICF) of a correct response for item parameters under the 1PL model at Latent1D point. It updates the gradient vectors if they are not empty. It follows the parametrization $a(θ - b)$.

Arguments

latent::Latent1D : Required. A 1-dimensional Latent1D latent variable.

parameters::Parameters1PL : Required. A 1-parameter logistic parameters object.

Output

A Float64 scalar.

Psychometrics._probability — Method

_probability(latent::Latent1D, parameters::Parameters1PL)

Description

It computes the probability (ICF) of a correct response for item parameters under the 1PL model at Latent1D point. It follows the parametrization $a(θ - b)$.

Arguments

latent::Latent1D : Required. A 1-dimensional Latent1D latent variable.

parameters::Parameters1PL : Required. A 1-parameter logistic parameters object.

Output

A Float64 scalar.

Psychometrics._probability — Method

_probability(latent::Latent1D, parameters::Parameters2PL, g_item::Vector{Float64}, g_latent::Vector{Float64})

Description

It computes the probability (ICF) of a correct response for item parameters under the 2PL model at Latent1D point. It updates the gradient vectors if they are not empty. It follows the parametrization $a(θ - b)$.

Arguments

latent::Latent1D : Required. A 1-dimensional Latent1D latent variable.

parameters::Parameters2PL : Required. A 2-parameter logistic parameters object.

Output

A Float64 scalar.

Psychometrics._probability — Method

_probability(latent::Latent1D, parameters::Parameters2PL)

Description

It computes the probability (ICF) of a correct response for item parameters under the 2PL model at Latent1D point. It follows the parametrization $a(θ - b)$.

Arguments

latent::Latent1D : Required. A 1-dimensional Latent1D latent variable.

parameters::Parameters2PL : Required. A 2-parameter logistic parameters object.

Output

A Float64 scalar.

Psychometrics._probability — Method

_probability(latent::Latent1D, parameters::Parameters3PL,  g_item::Vector{Float64}, g_latent::Vector{Float64})

Description

It computes the probability (ICF) of a correct response for item parameters under the 3PL model at Latent1D point. It updates the gradient vectors if they are not empty. It follows the parametrization $a(θ - b)$.

Arguments

latent::Latent1D : Required. A 1-dimensional Latent1D latent variable.

parameters::Parameters3PL : Required. A 3-parameter logistic parameters object.

Output

A Float64 scalar.

Psychometrics._probability — Method

_probability(latent::Latent1D, parameters::Parameters3PL)

Description

It computes the probability (ICF) of a correct response for item parameters under the 3PL model at Latent1D point. It follows the parametrization $a(θ - b)$.

Arguments

latent::Latent1D : Required. A 1-dimensional Latent1D latent variable.

parameters::Parameters3PL : Required. A 3-parameter logistic parameters object.

Output

A Float64 scalar.

Psychometrics._probability — Method

_probability(latent::Latent1D, parameters::Parameters3PL,  g_item::Vector{Float64}, g_latent::Vector{Float64})

Description

It computes the probability (ICF) of a correct response for item parameters under the 3PL model at Latent1D point. It updates the gradient vectors if they are not empty. It follows the parametrization $a(θ - b)$.

Arguments

latent::LatentND : Required. A N-dimensional latent variable.

parameters::ParametersNPL : Required. A 2-parameter logistic parameters object suitable for a N-dimensional latent.

Output

A Float64 scalar.

Psychometrics.probability — Method

probability(examinee::AbstractExaminee, item::AbstractItem)

Description

It computes the probability (ICF) that an examinee answers correctly at item.

Arguments

examinee::AbstractExaminee : Required. An Examinee.

item::AbstractItem : Required. An Item.

Output

A Float64 scalar.

Psychometrics.probability — Method

probability(parameters_matrix::Matrix{Float64}, latent_matrix::Matrix{Float64})

Description

It computes the probability (ICF) of a correct response for item parameters and latents values provided in matrix form. Not suitable for 3PL models, for such a kind of model use probability_3PL(). It follows the parametrization $aθ - b$.

Arguments

parameters_matrix::Matrix{Float64} : Required. A n_latents x N matrix with latents values.

latent_matrix::Matrix{Float64} : Required. A I x (n_latents + 1) matrix with item parameters. intercept (b) must be in first column, latents coefficients (aj) in next columns (2, ..., nlatents + 1).

Output

A I x N Float64 matrix.

Psychometrics.probability — Method

probability(examinees::Vector{<:AbstractExaminee}, items::Vector{<:AbstractItem})

Description

It computes the probability (ICF) that a vector of examinees answers correctly at a vector of items.

Arguments

examinees::Vector{<:AbstractExaminee} : Required. A vector of Examinees.

items::Vector{<:AbstractItem} : Required. A vector of Items.

Output

A matrixexa.

Psychometrics.probability — Method

probability(items::Vector{<:AbstractItem}, examinees::Vector{<:AbstractExaminee})

Description

It computes the probability (ICF) that a vector of examinees answers correctly at a vector of items.

Arguments

examinees::Vector{<:AbstractExaminee} : Required. A vector of Examinees.

items::Vector{<:AbstractItem} : Required. A vector of Items.

Output

A matrix.

Psychometrics.probability_3PL — Method

probability_3PL(parameters_matrix::Matrix{Float64}, latent_matrix::Matrix{Float64})

Description

Only for models which has guessing parameter (c) in last row of parameters_matrix. It computes the probability (ICF) of a correct response for item parameters and latents values provided in matrix form. It follows the parametrization $aθ - b$.

Arguments

parameters_matrix::Matrix{Float64} : Required. A n_latents x N matrix with latents values.

latent_matrix::Matrix{Float64} : Required. A I x (n_latents + 1) matrix with item parameters. intercept (b) must be in first column, latents coefficients (aj) in next columns (2, ..., nlatents + 1).

Output

A I x N Float64 matrix.