Probability (Item Characteristic Function)

Psychometrics.probability — Method

 probability(latent_val::Float64, parameters::Parameters1PL)

Description

It computes the probability (ICF) of a correct response for item parameters under the 1PL model at latent_val point.

Arguments

latent_val::Float64 : Required. The point in the latent space in which compute the probability.

parameters::Parameters1PL : Required. A 1-parameter logistic parameters object.

Output

A Float64 scalar.

Psychometrics.probability — Method

 probability(latent_val::Float64, parameters::Parameters2PL)

Description

It computes the probability (ICF) of a correct response for item parameters under the 2PL model at latent_val point.

Arguments

latent_val::Float64 : Required. The point in the latent space in which compute the probability.

parameters::Parameters2PL : Required. A 2-parameter logistic parameters object.

Output

A Float64 scalar.

Psychometrics.probability — Method

probability(latent_val::Float64, parameters::Parameters3PL)

Description

It computes the probability (ICF) of a correct response for item parameters under the 3PL model at latent_val point.

Arguments

latent_val::Float64 : Required. The point in the latent space in which compute the probability.

parameters::Parameters3PL : Required. A 3-parameter logistic parameters object.

Output

A Float64 scalar.

Psychometrics.probability — Method

probability(latent::Latent1D, parameters::Parameters1PL, g_item::Vector{Float64}, g_latent::Vector{Float64})

Description

It computes the probability (ICF) of a correct response for item parameters under the 1PL model at Latent1D point. It updates the gradient vectors if they are not empty.

Arguments

latent::Latent1D : Required. A 1-dimensional Latent1D latent variable.

parameters::Parameters1PL : Required. A 1-parameter logistic parameters object.

Output

A Float64 scalar.

Psychometrics.probability — Method

 probability(latent::Latent1D, parameters::Parameters1PL)

Description

It computes the probability (ICF) of a correct response for item parameters under the 1PL model at Latent1D point.

Arguments

latent::Latent1D : Required. A 1-dimensional Latent1D latent variable.

parameters::Parameters1PL : Required. A 1-parameter logistic parameters object.

Output

A Float64 scalar.

Psychometrics.probability — Method

probability(latent::Latent1D, parameters::Parameters2PL, g_item::Vector{Float64}, g_latent::Vector{Float64})

Description

It computes the probability (ICF) of a correct response for item parameters under the 2PL model at Latent1D point. It updates the gradient vectors if they are not empty.

Arguments

latent::Latent1D : Required. A 1-dimensional Latent1D latent variable.

parameters::Parameters2PL : Required. A 2-parameter logistic parameters object.

Output

A Float64 scalar.

Psychometrics.probability — Method

probability(latent::Latent1D, parameters::Parameters2PL)

Description

It computes the probability (ICF) of a correct response for item parameters under the 2PL model at Latent1D point.

Arguments

latent::Latent1D : Required. A 1-dimensional Latent1D latent variable.

parameters::Parameters2PL : Required. A 2-parameter logistic parameters object.

Output

A Float64 scalar.

Psychometrics.probability — Method

probability(latent::Latent1D, parameters::Parameters3PL,  g_item::Vector{Float64}, g_latent::Vector{Float64})

Description

It computes the probability (ICF) of a correct response for item parameters under the 3PL model at Latent1D point. It updates the gradient vectors if they are not empty.

Arguments

latent::Latent1D : Required. A 1-dimensional Latent1D latent variable.

parameters::Parameters3PL : Required. A 3-parameter logistic parameters object.

Output

A Float64 scalar.

Psychometrics.probability — Method

probability(latent::Latent1D, parameters::Parameters3PL)

Description

It computes the probability (ICF) of a correct response for item parameters under the 3PL model at Latent1D point.

Arguments

latent::Latent1D : Required. A 1-dimensional Latent1D latent variable.

parameters::Parameters3PL : Required. A 3-parameter logistic parameters object.

Output

A Float64 scalar.