Response Structs and Methods

Psychometrics.add_response! — Method

add_response!(response::AbstractResponse, responses::Vector{Response})

Push the response in the response vector responses.

Psychometrics.generate_response — Method

generate_response(examinee::AbstractExaminee, item::AbstractItem)

Randomly generate a response by examinee to item.

Psychometrics.generate_response — Method

generate_response(latent::Latent1D, parameters::AbstractParameters)

Randomly generate a response for a 1-dimensional latent variable and custom item parameters.

Psychometrics.get_examinee_responses — Method

get_examinee_responses(idx::Int64, responses::Vector{Response})

It returns the vector of responses given by examinee with index idx.

Psychometrics.get_examinees — Method

get_examinees(item_idx::Int64, responses::Vector{<:AbstractResponse})

It returns the esaminees who answered to the item with index item_idx.

Psychometrics.get_item_responses — Method

get_item_responses(idx::Int64, responses::Vector{Response})

It returns the vector of responses given to item with index idx.

Psychometrics.get_items — Method

get_items(examinee_idx::Int64, responses::Vector{<:AbstractResponse})

It returns the items answered by the examinee with index examinee_idx.