Responses

Psychometrics.add_response! — Method

add_response!(response::AbstractResponse, responses::Vector{Response})

Push the response in the response vector responses.

Psychometrics.answer — Method

answer(Examinee::AbstractExaminee, item::AbstractParameters)

Randomly generate a response by Examinee to item.

Psychometrics.answer — Method

answer(examinees:AbstractExaminee}, items::Vector{<:AbstractItem})

Randomly generate responses by all the examinees in examinees to items in items.

Psychometrics.answer — Method

answer(examinee_id::String, item_id::String, examinees::Vector{<:AbstractExaminee}, items::Vector{<:AbstractItem})

Randomly generate a response by Examinee with index examinee_id to item with index item_id.

Psychometrics.generate_response — Method

generate_response(examinee::AbstractExaminee, item::AbstractItem)

Randomly generate a response by examinee to item.

Psychometrics.generate_response — Method

generate_response(examinee::AbstractExaminee, items::Vector{<:AbstractItem})

Randomly generate a response by examinee to items.

Psychometrics.generate_response — Method

generate_response(examinees::Vector{<:AbstractExaminee}, items::Vector{<:AbstractItem}, design::Vector{Vector{Int64}})

Randomly generate a response by examinee to item under a design dictionary of the same size of examinees with indices of items which must be answered by each examinee.

Psychometrics.generate_response — Method

generate_response(latent::Latent1D, parameters::AbstractParameters)

Randomly generate a response for a 1-dimensional latent variable and custom item parameters.

Psychometrics.get_design_matrix — Method

get_design_matrix(responses::Vector{<:AbstractResponse}, I::Int64, N::Int64)

Returns the I x N design matrix.

Psychometrics.get_examinees_by_item_id — Method

get_examinees_by_item_id(item_id::String, responses::Vector{<:AbstractResponse}, examinees::Vector{<:AbstractExaminee})

It returns the examinees who answered to the item with id item_id.

Psychometrics.get_items_by_examinee_id — Method

get_items_by_examinee_id(examinee_id::String, responses::Vector{<:AbstractItem})

It returns the items answered by the examinee with id examinee_id.

Psychometrics.get_response_matrix — Method

get_response_matrix(responses::Vector{<:AbstractResponse}, I::Int64, N::Int64)

Transform vector of Responses in a I x N response matrix. A non given answer has value 0.0.

Psychometrics.get_responses — Method

get_responses(response_matrix::Matrix{Float64}, design_matrix::Matrix{Float64}, items::Vector{<:AbstractItem}, examinees::Vector{<:AbstractExaminee})

Transforms a I x N response matrix in a vector of Responses given a valid design_matrix, a vector of Items and a vector of Examinees.

Psychometrics.get_responses_by_examinee_id — Method

get_responses_by_examinee_id(examinee_id::String, responses::Vector{Response})

It returns the vector of responses given by examinee with id = id.

Psychometrics.get_responses_by_item_id — Method

get_responses_by_item_id(item_id::String, responses::Vector{Response})

It returns the vector of responses given to item with id = id.